Split emit_wasm/calls.rs (4609 lines) into per-module files and add stdlib_dispatch
- calls.rs 4609 -> 1144 lines: extracted per-stdlib-module dispatch into dedicated files
  - calls_env.rs (env), calls_random.rs (random), calls_datetime.rs (datetime + decimal helpers)
  - calls_http.rs (http), calls_fs.rs (fs + path/filetype/list-dir helpers)
  - calls_io.rs (io), calls_process.rs (process/stdin_lines)
- New stdlib_dispatch.rs: StdlibOp enum (Call1/Call2/Call3/FloatUnaryCall) for declarative emit of simple runtime-call patterns
- Migrate math/float/string simple ops to the declarative dispatcher
- Register test-function names with __test_ prefix at module level to avoid colliding with user functions of the same name (e.g. fn broadcast_add + test "broadcast_add")
- Suppress ALMIDE_WASM_STUB_WARN output by default; ALMIDE_WASM_STUB_PANIC=1 fails the compile when a stub_call is reached
