+++ crates/almide-codegen/src/emit_wasm/calls.rs
@@ -485,7 +485,14 @@ impl FuncCompiler<'_> {
                                 for arg in args { self.emit_expr(arg); }
                                 wasm!(self.func, { call(func_idx); });
                             } else {
-                                self.emit_stub_call(args);
+                                // Last resort: bare func name (for cross-module calls where
+                                // module name differs from canonical)
+                                if let Some(&func_idx) = self.emitter.func_map.get(func.as_str()) {
+                                    for arg in args { self.emit_expr(arg); }
+                                    wasm!(self.func, { call(func_idx); });
+                                } else {
+                                    self.emit_stub_call(args);
+                                }
                             }
                         }
                     }
@@ -782,11 +789,21 @@ impl FuncCompiler<'_> {
 
     pub(super) fn emit_stub_call(&mut self, args: &[IrExpr]) {
         // Unimplemented function: trap immediately rather than returning a default value.
-        // Returning silent defaults (0, empty string, etc.) is dangerous in medical contexts
-        // where incorrect results could go unnoticed.
-        if std::env::var("ALMIDE_WASM_STUB_WARN").is_ok() {
-            eprintln!("[WASM] stub call reached — will trap at runtime");
+        // After ResolveCalls pass, user-module unresolved calls are a compile error —
+        // so any stub_call reaching here must be an unimplemented stdlib method.
+        //
+        // DEFAULT: warn + continue (emit `unreachable` WASM — runtime trap).
+        // ALMIDE_WASM_STUB_PANIC=1: panic at compile time (hard assertion for CI).
+        //
+        // Today's audit confirmed all 183 spec + 60 nn WASM tests compile
+        // without hitting any stub_call, so in practice this is dead code.
+        // We keep the branch to have a graceful fallback for future stdlib
+        // additions that haven't got WASM impls yet.
+        if std::env::var("ALMIDE_WASM_STUB_PANIC").is_ok() {
+            panic!("[WASM STUB] stub_call reached — compile-time hard failure (ALMIDE_WASM_STUB_PANIC=1)");
         }
+        eprintln!("[WASM STUB] stub_call emitted — will trap at runtime. \
+                   Set ALMIDE_WASM_STUB_PANIC=1 to fail compile instead.");
         for arg in args {
             self.emit_expr(arg);
             if values::ty_to_valtype(&arg.ty).is_some() {
@@ -910,7 +927,7 @@ impl FuncCompiler<'_> {
             "map" => {
                 // fan.map(xs, f) — apply effect fn f to each element, unwrap Results
                 // f returns Result[T, E] (i32 ptr). We unwrap ok values into result list.
-                let elem_ty = self.list_elem_ty(&args[0].ty);
+                let elem_ty = self.resolve_list_elem(&args[0], None);
                 let es = values::byte_size(&elem_ty) as i32;
                 // Determine output element type from result_ty
                 let out_elem_ty = if let Ty::Applied(_, a) = result_ty {
@@ -1124,3469 +1141,4 @@ impl FuncCompiler<'_> {
         }
     }
 
-    /// Env module: environment access.
-    fn emit_env_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "args" => {
-                // env.args() → List[String]: return empty list (WASI args not implemented yet)
-                let s = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(s);
-                    local_get(s); i32_const(0); i32_store(0);
-                    local_get(s);
-                });
-                self.scratch.free_i32(s);
-            }
-            "unix_timestamp" => {
-                // WASI clock_time_get(id=0 realtime, precision=0, time_ptr)
-                // Returns nanoseconds as i64, convert to seconds
-                let time_ptr = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(time_ptr);
-                    i32_const(0); // clock_id: realtime
-                    i64_const(0); // precision
-                    local_get(time_ptr);
-                    call(self.emitter.rt.clock_time_get);
-                    drop; // discard error code
-                    local_get(time_ptr); i64_load(0);
-                    i64_const(1000000000); i64_div_u;
-                });
-                self.scratch.free_i32(time_ptr);
-            }
-            "millis" => {
-                // WASI clock_time_get(id=0 realtime, precision=0, time_ptr)
-                // Returns nanoseconds as i64, convert to milliseconds
-                let time_ptr = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(time_ptr);
-                    i32_const(0); // clock_id: realtime
-                    i64_const(0); // precision
-                    local_get(time_ptr);
-                    call(self.emitter.rt.clock_time_get);
-                    drop; // discard error code
-                    local_get(time_ptr); i64_load(0);
-                    i64_const(1000000); i64_div_u;
-                });
-                self.scratch.free_i32(time_ptr);
-            }
-            "os" => {
-                let s = self.emitter.intern_string("wasi");
-                wasm!(self.func, { i32_const(s as i32); });
-            }
-            "temp_dir" => {
-                let s = self.emitter.intern_string("/tmp");
-                wasm!(self.func, { i32_const(s as i32); });
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
-
-    /// Log module: structured logging to stderr via WASI fd_write(2, ...).
-    /// Random module: PRNG-based random number generation.
-    /// Uses xorshift64 state stored at linear memory address 0 (8 bytes).
-    pub(super) fn emit_random_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "int" => {
-                // random.int(min, max) → Int in [min, max]
-                let min = self.scratch.alloc_i64();
-                let max = self.scratch.alloc_i64();
-                let state = self.scratch.alloc_i64();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(min); });
-                self.emit_expr(&args[1]);
-                wasm!(self.func, {
-                    local_set(max);
-                    // Load PRNG state from mem[0..8]
-                    i32_const(0); i64_load(0); local_set(state);
-                    // If state == 0, initialize with seed
-                    local_get(state); i64_eqz;
-                    if_empty;
-                      i32_const(0); i32_const(8); call(self.emitter.rt.random_get); drop;
-                      i32_const(0); i64_load(0); local_set(state);
-                      local_get(state); i64_eqz;
-                      if_empty; i64_const(1); local_set(state); end;
-                    end;
-                    // xorshift64
-                    local_get(state); local_get(state); i64_const(13); i64_shl; i64_xor; local_set(state);
-                    local_get(state); local_get(state); i64_const(7); i64_shr_u; i64_xor; local_set(state);
-                    local_get(state); local_get(state); i64_const(17); i64_shl; i64_xor; local_set(state);
-                    // Store back
-                    i32_const(0); local_get(state); i64_store(0);
-                    // result = min + abs(state) % (max - min + 1)
-                    local_get(min);
-                    // abs(state)
-                    local_get(state); i64_const(0); i64_lt_s;
-                    if_i64; i64_const(0); local_get(state); i64_sub; else_; local_get(state); end;
-                    local_get(max); local_get(min); i64_sub; i64_const(1); i64_add;
-                    i64_rem_u;
-                    i64_add;
-                });
-                self.scratch.free_i64(state);
-                self.scratch.free_i64(max);
-                self.scratch.free_i64(min);
-            }
-            "float" => {
-                // random.float() → Float in [0.0, 1.0)
-                let state = self.scratch.alloc_i64();
-                wasm!(self.func, {
-                    i32_const(0); i64_load(0); local_set(state);
-                    local_get(state); i64_eqz;
-                    if_empty;
-                      i32_const(0); i32_const(8); call(self.emitter.rt.random_get); drop;
-                      i32_const(0); i64_load(0); local_set(state);
-                      local_get(state); i64_eqz;
-                      if_empty; i64_const(1); local_set(state); end;
-                    end;
-                    local_get(state); local_get(state); i64_const(13); i64_shl; i64_xor; local_set(state);
-                    local_get(state); local_get(state); i64_const(7); i64_shr_u; i64_xor; local_set(state);
-                    local_get(state); local_get(state); i64_const(17); i64_shl; i64_xor; local_set(state);
-                    i32_const(0); local_get(state); i64_store(0);
-                });
-                // Convert to float in [0, 1): abs(state) >>> 11 * (1/2^53)
-                wasm!(self.func, {
-                    local_get(state); i64_const(0); i64_lt_s;
-                    if_i64; i64_const(0); local_get(state); i64_sub; else_; local_get(state); end;
-                    i64_const(11); i64_shr_u;
-                    f64_convert_i64_u;
-                    f64_const(1.1102230246251565e-16); // 1.0 / 2^53
-                    f64_mul;
-                });
-                self.scratch.free_i64(state);
-            }
-            "choice" => {
-                // random.choice(xs: List[T]) → Option[T]: none if empty, some(random elem) if non-empty
-                let xs = self.scratch.alloc_i32();
-                let idx = self.scratch.alloc_i32();
-                let option_box = self.scratch.alloc_i32();
-                let state = self.scratch.alloc_i64();
-                let elem_ty = self.list_elem_ty(&args[0].ty);
-                let es = values::byte_size(&elem_ty) as i32;
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    local_set(xs);
-                    local_get(xs); i32_load(0); i32_eqz;
-                    if_i32;
-                      i32_const(0); // none (empty list)
-                    else_;
-                      // PRNG to get random index
-                      i32_const(0); i64_load(0); local_set(state);
-                      local_get(state); i64_eqz;
-                      if_empty;
-                      i32_const(0); i32_const(8); call(self.emitter.rt.random_get); drop;
-                      i32_const(0); i64_load(0); local_set(state);
-                      local_get(state); i64_eqz;
-                      if_empty; i64_const(1); local_set(state); end;
-                    end;
-                      local_get(state); local_get(state); i64_const(13); i64_shl; i64_xor; local_set(state);
-                      local_get(state); local_get(state); i64_const(7); i64_shr_u; i64_xor; local_set(state);
-                      local_get(state); local_get(state); i64_const(17); i64_shl; i64_xor; local_set(state);
-                      i32_const(0); local_get(state); i64_store(0);
-                      // idx = abs(state) % len
-                      local_get(state); i64_const(0); i64_lt_s;
-                      if_i64; i64_const(0); local_get(state); i64_sub; else_; local_get(state); end;
-                      local_get(xs); i32_load(0); i64_extend_i32_u;
-                      i64_rem_u; i32_wrap_i64; local_set(idx);
-                      // Alloc option box and load elem into it
-                      i32_const(es); call(self.emitter.rt.alloc); local_set(option_box);
-                      local_get(option_box);
-                      local_get(xs); i32_const(4); i32_add;
-                      local_get(idx); i32_const(es); i32_mul; i32_add;
-                });
-                self.emit_load_at(&elem_ty, 0);
-                self.emit_store_at(&elem_ty, 0);
-                wasm!(self.func, {
-                      local_get(option_box);
-                    end;
-                });
-                self.scratch.free_i64(state);
-                self.scratch.free_i32(option_box);
-                self.scratch.free_i32(idx);
-                self.scratch.free_i32(xs);
-            }
-            "shuffle" => {
-                // random.shuffle(xs) → List[T]: Fisher-Yates shuffle on a copy
-                let elem_ty = self.list_elem_ty(&args[0].ty);
-                let es = values::byte_size(&elem_ty) as i32;
-                let src = self.scratch.alloc_i32();
-                let dst = self.scratch.alloc_i32();
-                let len = self.scratch.alloc_i32();
-                let i = self.scratch.alloc_i32();
-                let j = self.scratch.alloc_i32();
-                let state = self.scratch.alloc_i64();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    local_set(src);
-                    local_get(src); i32_load(0); local_set(len);
-                    // Alloc copy
-                    i32_const(4); local_get(len); i32_const(es); i32_mul; i32_add;
-                    call(self.emitter.rt.alloc); local_set(dst);
-                    local_get(dst); local_get(len); i32_store(0);
-                    // Copy all elements
-                    i32_const(0); local_set(i);
-                    block_empty; loop_empty;
-                      local_get(i); local_get(len); i32_ge_u; br_if(1);
-                      local_get(dst); i32_const(4); i32_add;
-                      local_get(i); i32_const(es); i32_mul; i32_add;
-                      local_get(src); i32_const(4); i32_add;
-                      local_get(i); i32_const(es); i32_mul; i32_add;
-                });
-                self.emit_elem_copy(&elem_ty);
-                wasm!(self.func, {
-                      local_get(i); i32_const(1); i32_add; local_set(i);
-                      br(0);
-                    end; end;
-                    // Fisher-Yates shuffle (backwards)
-                    local_get(len); i32_const(1); i32_sub; local_set(i);
-                    i32_const(0); i64_load(0); local_set(state);
-                    local_get(state); i64_eqz;
-                    if_empty;
-                      i32_const(0); i32_const(8); call(self.emitter.rt.random_get); drop;
-                      i32_const(0); i64_load(0); local_set(state);
-                      local_get(state); i64_eqz;
-                      if_empty; i64_const(1); local_set(state); end;
-                    end;
-                    block_empty; loop_empty;
-                      local_get(i); i32_const(0); i32_le_s; br_if(1);
-                      // xorshift
-                      local_get(state); local_get(state); i64_const(13); i64_shl; i64_xor; local_set(state);
-                      local_get(state); local_get(state); i64_const(7); i64_shr_u; i64_xor; local_set(state);
-                      local_get(state); local_get(state); i64_const(17); i64_shl; i64_xor; local_set(state);
-                      // j = abs(state) % (i + 1)
-                      local_get(state);
-                      local_get(state); i64_const(0); i64_lt_s;
-                      if_i64; i64_const(0); local_get(state); i64_sub; else_; local_get(state); end;
-                      local_get(i); i32_const(1); i32_add; i64_extend_i32_u;
-                      i64_rem_u; i32_wrap_i64; local_set(j);
-                      // swap dst[i] and dst[j] using mem[0..es] as temp
-                      // Copy dst[i] to temp
-                      i32_const(0);
-                      local_get(dst); i32_const(4); i32_add;
-                      local_get(i); i32_const(es); i32_mul; i32_add;
-                      i32_load(0); i32_store(0);
-                      // Copy dst[j] to dst[i]
-                      local_get(dst); i32_const(4); i32_add;
-                      local_get(i); i32_const(es); i32_mul; i32_add;
-                      local_get(dst); i32_const(4); i32_add;
-                      local_get(j); i32_const(es); i32_mul; i32_add;
-                      i32_load(0); i32_store(0);
-                      // Copy temp to dst[j]
-                      local_get(dst); i32_const(4); i32_add;
-                      local_get(j); i32_const(es); i32_mul; i32_add;
-                      i32_const(0); i32_load(0); i32_store(0);
-                      local_get(i); i32_const(1); i32_sub; local_set(i);
-                      br(0);
-                    end; end;
-                    i32_const(0); local_get(state); i64_store(0); // save state
-                    local_get(dst);
-                });
-                self.scratch.free_i64(state);
-                self.scratch.free_i32(j);
-                self.scratch.free_i32(i);
-                self.scratch.free_i32(len);
-                self.scratch.free_i32(dst);
-                self.scratch.free_i32(src);
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
-
-    /// datetime module: all functions operate on i64 Unix timestamps (seconds since 1970-01-01 UTC).
-    /// Uses the proleptic Gregorian calendar via Julian Day Number conversions.
-    pub(super) fn emit_datetime_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "from_parts" => {
-                // datetime.from_parts(year, month, day, hour, minute, second) → Int
-                // JDN algorithm: a=(14-month)/12, y=year+4800-a, m=month+12*a-3
-                // jdn = day + (153*m+2)/5 + 365*y + y/4 - y/100 + y/400 - 32045
-                // timestamp = (jdn - 2440588) * 86400 + h*3600 + min*60 + sec
-                let year = self.scratch.alloc_i64();
-                let month = self.scratch.alloc_i64();
-                let day = self.scratch.alloc_i64();
-                let hour = self.scratch.alloc_i64();
-                let minute = self.scratch.alloc_i64();
-                let second = self.scratch.alloc_i64();
-                let a = self.scratch.alloc_i64();
-                let y = self.scratch.alloc_i64();
-                let m = self.scratch.alloc_i64();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(year); });
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(month); });
-                self.emit_expr(&args[2]);
-                wasm!(self.func, { local_set(day); });
-                self.emit_expr(&args[3]);
-                wasm!(self.func, { local_set(hour); });
-                self.emit_expr(&args[4]);
-                wasm!(self.func, { local_set(minute); });
-                self.emit_expr(&args[5]);
-                wasm!(self.func, { local_set(second); });
-
-                wasm!(self.func, {
-                    i64_const(14); local_get(month); i64_sub; i64_const(12); i64_div_s; local_set(a);
-                    local_get(year); i64_const(4800); i64_add; local_get(a); i64_sub; local_set(y);
-                    local_get(month); i64_const(12); local_get(a); i64_mul; i64_add; i64_const(3); i64_sub; local_set(m);
-                    local_get(day);
-                    i64_const(153); local_get(m); i64_mul; i64_const(2); i64_add; i64_const(5); i64_div_s;
-                    i64_add;
-                    i64_const(365); local_get(y); i64_mul;
-                    i64_add;
-                    local_get(y); i64_const(4); i64_div_s;
-                    i64_add;
-                    local_get(y); i64_const(100); i64_div_s;
-                    i64_sub;
-                    local_get(y); i64_const(400); i64_div_s;
-                    i64_add;
-                    i64_const(32045); i64_sub;
-                    i64_const(2440588); i64_sub;
-                    i64_const(86400); i64_mul;
-                    local_get(hour); i64_const(3600); i64_mul; i64_add;
-                    local_get(minute); i64_const(60); i64_mul; i64_add;
-                    local_get(second); i64_add;
-                });
-
-                self.scratch.free_i64(m);
-                self.scratch.free_i64(y);
-                self.scratch.free_i64(a);
-                self.scratch.free_i64(second);
-                self.scratch.free_i64(minute);
-                self.scratch.free_i64(hour);
-                self.scratch.free_i64(day);
-                self.scratch.free_i64(month);
-                self.scratch.free_i64(year);
-            }
-            "year" | "month" | "day" => {
-                // Inverse JDN algorithm to extract date component from timestamp.
-                let ts = self.scratch.alloc_i64();
-                let d = self.scratch.alloc_i64();
-                let f = self.scratch.alloc_i64();
-                let e = self.scratch.alloc_i64();
-                let g = self.scratch.alloc_i64();
-                let h = self.scratch.alloc_i64();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    local_set(ts);
-                    // floor(ts / 86400)
-                    local_get(ts); i64_const(0); i64_ge_s;
-                    if_i64;
-                      local_get(ts); i64_const(86400); i64_div_s;
-                    else_;
-                      local_get(ts); i64_const(86399); i64_sub; i64_const(86400); i64_div_s;
-                    end;
-                    local_set(d);
-                    local_get(d); i64_const(2440588); i64_add; local_set(d);
-                    local_get(d); i64_const(1401); i64_add;
-                    i64_const(4); local_get(d); i64_mul; i64_const(274277); i64_add;
-                    i64_const(146097); i64_div_s; i64_const(3); i64_mul; i64_const(4); i64_div_s;
-                    i64_add; i64_const(38); i64_sub;
-                    local_set(f);
-                    i64_const(4); local_get(f); i64_mul; i64_const(3); i64_add; local_set(e);
-                    local_get(e); i64_const(1461); i64_rem_s; i64_const(4); i64_div_s; local_set(g);
-                    i64_const(5); local_get(g); i64_mul; i64_const(2); i64_add; local_set(h);
-                });
-
-                match func {
-                    "day" => {
-                        wasm!(self.func, {
-                            local_get(h); i64_const(153); i64_rem_s; i64_const(5); i64_div_s; i64_const(1); i64_add;
-                        });
-                    }
-                    "month" => {
-                        wasm!(self.func, {
-                            local_get(h); i64_const(153); i64_div_s; i64_const(2); i64_add;
-                            i64_const(12); i64_rem_s; i64_const(1); i64_add;
-                        });
-                    }
-                    "year" => {
-                        let mm = self.scratch.alloc_i64();
-                        wasm!(self.func, {
-                            local_get(h); i64_const(153); i64_div_s; i64_const(2); i64_add;
-                            i64_const(12); i64_rem_s; i64_const(1); i64_add;
-                            local_set(mm);
-                            local_get(e); i64_const(1461); i64_div_s; i64_const(4716); i64_sub;
-                            i64_const(14); local_get(mm); i64_sub; i64_const(12); i64_div_s;
-                            i64_add;
-                        });
-                        self.scratch.free_i64(mm);
-                    }
-                    _ => unreachable!(),
-                }
-
-                self.scratch.free_i64(h);
-                self.scratch.free_i64(g);
-                self.scratch.free_i64(e);
-                self.scratch.free_i64(f);
-                self.scratch.free_i64(d);
-                self.scratch.free_i64(ts);
-            }
-            "hour" => {
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    i64_const(86400); i64_rem_s;
-                    i64_const(86400); i64_add; i64_const(86400); i64_rem_s;
-                    i64_const(3600); i64_div_s;
-                });
-            }
-            "minute" => {
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    i64_const(3600); i64_rem_s;
-                    i64_const(3600); i64_add; i64_const(3600); i64_rem_s;
-                    i64_const(60); i64_div_s;
-                });
-            }
-            "second" => {
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    i64_const(60); i64_rem_s;
-                    i64_const(60); i64_add; i64_const(60); i64_rem_s;
-                });
-            }
-            "now" => {
-                // Call WASI clock_time_get(id=0 realtime, precision=0, time_ptr)
-                // Returns nanoseconds as i64 at time_ptr, convert to seconds
-                let time_ptr = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    // Allocate 8 bytes for i64 result (allocator guarantees 8-byte alignment)
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(time_ptr);
-                    // clock_time_get(id=0, precision=0, time_ptr)
-                    i32_const(0); // clock_id: realtime
-                    i64_const(0); // precision
-                    local_get(time_ptr); // output pointer (8-byte aligned)
-                    call(self.emitter.rt.clock_time_get);
-                    drop; // discard error code
-                    // Load i64 nanoseconds, convert to seconds
-                    local_get(time_ptr); i64_load(0);
-                    i64_const(1000000000); i64_div_u;
-                });
-                self.scratch.free_i32(time_ptr);
-            }
-            "add_days" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_const(86400); i64_mul; i64_add; });
-            }
-            "add_hours" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_const(3600); i64_mul; i64_add; });
-            }
-            "add_minutes" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_const(60); i64_mul; i64_add; });
-            }
-            "add_seconds" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_add; });
-            }
-            "from_unix" | "to_unix" => {
-                self.emit_expr(&args[0]);
-            }
-            "diff_seconds" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_sub; });
-            }
-            "is_before" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_lt_s; });
-            }
-            "is_after" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_gt_s; });
-            }
-            "diff_days" => {
-                self.emit_expr(&args[0]);
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i64_sub; i64_const(86400); i64_div_s; });
-            }
-            "format" => {
-                // Stub: return int.to_string(ts), ignore fmt
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { call(self.emitter.rt.int_to_string); });
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { drop; });
-            }
-            "to_iso" => {
-                // datetime.to_iso(ts) → String "YYYY-MM-DDTHH:MM:SSZ"
-                let ts = self.scratch.alloc_i64();
-                let ptr = self.scratch.alloc_i32();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(ts); });
-
-                wasm!(self.func, {
-                    i32_const(24); call(self.emitter.rt.alloc); local_set(ptr);
-                    local_get(ptr); i32_const(20); i32_store(0);
-                });
-
-                let d = self.scratch.alloc_i64();
-                let f = self.scratch.alloc_i64();
-                let e = self.scratch.alloc_i64();
-                let g = self.scratch.alloc_i64();
-                let h = self.scratch.alloc_i64();
-                let yr = self.scratch.alloc_i64();
-                let mo = self.scratch.alloc_i64();
-                let dy = self.scratch.alloc_i64();
-                let hr = self.scratch.alloc_i64();
-                let mi = self.scratch.alloc_i64();
-                let se = self.scratch.alloc_i64();
-
-                wasm!(self.func, {
-                    local_get(ts); i64_const(0); i64_ge_s;
-                    if_i64;
-                      local_get(ts); i64_const(86400); i64_div_s;
-                    else_;
-                      local_get(ts); i64_const(86399); i64_sub; i64_const(86400); i64_div_s;
-                    end;
-                    local_set(d);
-                    local_get(d); i64_const(2440588); i64_add; local_set(d);
-                    local_get(d); i64_const(1401); i64_add;
-                    i64_const(4); local_get(d); i64_mul; i64_const(274277); i64_add;
-                    i64_const(146097); i64_div_s; i64_const(3); i64_mul; i64_const(4); i64_div_s;
-                    i64_add; i64_const(38); i64_sub; local_set(f);
-                    i64_const(4); local_get(f); i64_mul; i64_const(3); i64_add; local_set(e);
-                    local_get(e); i64_const(1461); i64_rem_s; i64_const(4); i64_div_s; local_set(g);
-                    i64_const(5); local_get(g); i64_mul; i64_const(2); i64_add; local_set(h);
-                    local_get(h); i64_const(153); i64_rem_s; i64_const(5); i64_div_s; i64_const(1); i64_add; local_set(dy);
-                    local_get(h); i64_const(153); i64_div_s; i64_const(2); i64_add;
-                    i64_const(12); i64_rem_s; i64_const(1); i64_add; local_set(mo);
-                    local_get(e); i64_const(1461); i64_div_s; i64_const(4716); i64_sub;
-                    i64_const(14); local_get(mo); i64_sub; i64_const(12); i64_div_s;
-                    i64_add; local_set(yr);
-                    local_get(ts); i64_const(86400); i64_rem_s; i64_const(86400); i64_add; i64_const(86400); i64_rem_s;
-                    local_set(d);
-                    local_get(d); i64_const(3600); i64_div_s; local_set(hr);
-                    local_get(d); i64_const(3600); i64_rem_s; i64_const(60); i64_div_s; local_set(mi);
-                    local_get(d); i64_const(60); i64_rem_s; local_set(se);
-                });
-
-                self.emit_write_decimal_digits(ptr, 4, yr, 4);
-                wasm!(self.func, { local_get(ptr); i32_const(45); i32_store8(8); });
-                self.emit_write_decimal_digits(ptr, 9, mo, 2);
-                wasm!(self.func, { local_get(ptr); i32_const(45); i32_store8(11); });
-                self.emit_write_decimal_digits(ptr, 12, dy, 2);
-                wasm!(self.func, { local_get(ptr); i32_const(84); i32_store8(14); });
-                self.emit_write_decimal_digits(ptr, 15, hr, 2);
-                wasm!(self.func, { local_get(ptr); i32_const(58); i32_store8(17); });
-                self.emit_write_decimal_digits(ptr, 18, mi, 2);
-                wasm!(self.func, { local_get(ptr); i32_const(58); i32_store8(20); });
-                self.emit_write_decimal_digits(ptr, 21, se, 2);
-                wasm!(self.func, { local_get(ptr); i32_const(90); i32_store8(23); });
-
-                wasm!(self.func, { local_get(ptr); });
-
-                self.scratch.free_i64(se);
-                self.scratch.free_i64(mi);
-                self.scratch.free_i64(hr);
-                self.scratch.free_i64(dy);
-                self.scratch.free_i64(mo);
-                self.scratch.free_i64(yr);
-                self.scratch.free_i64(h);
-                self.scratch.free_i64(g);
-                self.scratch.free_i64(e);
-                self.scratch.free_i64(f);
-                self.scratch.free_i64(d);
-                self.scratch.free_i32(ptr);
-                self.scratch.free_i64(ts);
-            }
-            "weekday" => {
-                // (floor(ts/86400) + 4) % 7: 0=Sun..6=Sat
-                let ts = self.scratch.alloc_i64();
-                let wd = self.scratch.alloc_i64();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    local_set(ts);
-                    local_get(ts); i64_const(0); i64_ge_s;
-                    if_i64;
-                      local_get(ts); i64_const(86400); i64_div_s;
-                    else_;
-                      local_get(ts); i64_const(86399); i64_sub; i64_const(86400); i64_div_s;
-                    end;
-                    i64_const(4); i64_add;
-                    i64_const(7); i64_rem_s;
-                    i64_const(7); i64_add; i64_const(7); i64_rem_s;
-                    local_set(wd);
-                });
-
-                let sun = self.emitter.intern_string("Sunday");
-                let mon = self.emitter.intern_string("Monday");
-                let tue = self.emitter.intern_string("Tuesday");
-                let wed = self.emitter.intern_string("Wednesday");
-                let thu = self.emitter.intern_string("Thursday");
-                let fri = self.emitter.intern_string("Friday");
-                let sat = self.emitter.intern_string("Saturday");
-
-                wasm!(self.func, {
-                    local_get(wd); i64_eqz;
-                    if_i32; i32_const(sun as i32);
-                    else_;
-                      local_get(wd); i64_const(1); i64_eq;
-                      if_i32; i32_const(mon as i32);
-                      else_;
-                        local_get(wd); i64_const(2); i64_eq;
-                        if_i32; i32_const(tue as i32);
-                        else_;
-                          local_get(wd); i64_const(3); i64_eq;
-                          if_i32; i32_const(wed as i32);
-                          else_;
-                            local_get(wd); i64_const(4); i64_eq;
-                            if_i32; i32_const(thu as i32);
-                            else_;
-                              local_get(wd); i64_const(5); i64_eq;
-                              if_i32; i32_const(fri as i32);
-                              else_;
-                                i32_const(sat as i32);
-                              end;
-                            end;
-                          end;
-                        end;
-                      end;
-                    end;
-                });
-
-                self.scratch.free_i64(wd);
-                self.scratch.free_i64(ts);
-            }
-            "parse_iso" => {
-                // datetime.parse_iso(s: String) → Result[Int, String]
-                let s = self.scratch.alloc_i32();
-                let result = self.scratch.alloc_i32();
-                let yr = self.scratch.alloc_i64();
-                let mo = self.scratch.alloc_i64();
-                let dy = self.scratch.alloc_i64();
-                let hr = self.scratch.alloc_i64();
-                let mi = self.scratch.alloc_i64();
-                let se = self.scratch.alloc_i64();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(s); });
-
-                let err_msg = self.emitter.intern_string("invalid datetime format");
-                wasm!(self.func, {
-                    local_get(s); i32_load(0); i32_const(19); i32_lt_u;
-                    if_i32;
-                      i32_const(8); call(self.emitter.rt.alloc); local_set(result);
-                      local_get(result); i32_const(1); i32_store(0);
-                      local_get(result); i32_const(err_msg as i32); i32_store(4);
-                      local_get(result);
-                    else_;
-                });
-
-                self.emit_parse_digits(s, 0, 4, yr);
-                self.emit_parse_digits(s, 5, 2, mo);
-                self.emit_parse_digits(s, 8, 2, dy);
-                self.emit_parse_digits(s, 11, 2, hr);
-                self.emit_parse_digits(s, 14, 2, mi);
-                self.emit_parse_digits(s, 17, 2, se);
-
-                let a = self.scratch.alloc_i64();
-                let y = self.scratch.alloc_i64();
-                let m = self.scratch.alloc_i64();
-                wasm!(self.func, {
-                    i64_const(14); local_get(mo); i64_sub; i64_const(12); i64_div_s; local_set(a);
-                    local_get(yr); i64_const(4800); i64_add; local_get(a); i64_sub; local_set(y);
-                    local_get(mo); i64_const(12); local_get(a); i64_mul; i64_add; i64_const(3); i64_sub; local_set(m);
-                    local_get(dy);
-                    i64_const(153); local_get(m); i64_mul; i64_const(2); i64_add; i64_const(5); i64_div_s; i64_add;
-                    i64_const(365); local_get(y); i64_mul; i64_add;
-                    local_get(y); i64_const(4); i64_div_s; i64_add;
-                    local_get(y); i64_const(100); i64_div_s; i64_sub;
-                    local_get(y); i64_const(400); i64_div_s; i64_add;
-                    i64_const(32045); i64_sub;
-                    i64_const(2440588); i64_sub;
-                    i64_const(86400); i64_mul;
-                    local_get(hr); i64_const(3600); i64_mul; i64_add;
-                    local_get(mi); i64_const(60); i64_mul; i64_add;
-                    local_get(se); i64_add;
-                    local_set(yr); // reuse as timestamp
-                    // Build ok Result: [tag=0:i32][timestamp:i64] = 12 bytes
-                    i32_const(12); call(self.emitter.rt.alloc); local_set(result);
-                    local_get(result); i32_const(0); i32_store(0);
-                    local_get(result); local_get(yr); i64_store(4);
-                    local_get(result);
-                    end;
-                });
-
-                self.scratch.free_i64(m);
-                self.scratch.free_i64(y);
-                self.scratch.free_i64(a);
-                self.scratch.free_i64(se);
-                self.scratch.free_i64(mi);
-                self.scratch.free_i64(hr);
-                self.scratch.free_i64(dy);
-                self.scratch.free_i64(mo);
-                self.scratch.free_i64(yr);
-                self.scratch.free_i32(result);
-                self.scratch.free_i32(s);
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
-
-    /// Write N decimal digits of an i64 value to a string buffer at a given byte offset.
-    fn emit_write_decimal_digits(&mut self, ptr: u32, byte_offset: u32, val: u32, num_digits: u32) {
-        let tmp = self.scratch.alloc_i64();
-        wasm!(self.func, { local_get(val); local_set(tmp); });
-        for i in (0..num_digits).rev() {
-            let off = byte_offset + i;
-            wasm!(self.func, {
-                local_get(ptr);
-                local_get(tmp); i64_const(10); i64_rem_s;
-                i64_const(48); i64_add;
-                i32_wrap_i64;
-                i32_store8(off);
-                local_get(tmp); i64_const(10); i64_div_s; local_set(tmp);
-            });
-        }
-        self.scratch.free_i64(tmp);
-    }
-
-    /// Parse N decimal ASCII digits from a string buffer into an i64 local.
-    fn emit_parse_digits(&mut self, str_local: u32, char_offset: u32, num_digits: u32, dest: u32) {
-        wasm!(self.func, { i64_const(0); local_set(dest); });
-        for i in 0..num_digits {
-            let off = 4 + char_offset + i;
-            wasm!(self.func, {
-                local_get(dest); i64_const(10); i64_mul;
-                local_get(str_local); i32_load8_u(off);
-                i64_extend_i32_u; i64_const(48); i64_sub;
-                i64_add;
-                local_set(dest);
-            });
-        }
-    }
-
-    /// HTTP module: response construction and header manipulation.
-    /// Response layout: [status:i64][body:i32 str ptr][headers:i32 list ptr]
-    /// Headers = List[(String, String)] — list of tuple pointers.
-    fn emit_http_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "response" => {
-                // http.response(status: Int, body: String) → Response
-                let s = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(s);
-                    local_get(s);
-                });
-                self.emit_expr(&args[0]); // status: i64
-                wasm!(self.func, { i64_store(0); local_get(s); });
-                self.emit_expr(&args[1]); // body: i32 str
-                wasm!(self.func, {
-                    i32_store(8);
-                    // Empty headers list
-                    local_get(s);
-                    i32_const(4); call(self.emitter.rt.alloc);
-                });
-                let empty_list = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    local_set(empty_list);
-                    local_get(empty_list); i32_const(0); i32_store(0);
-                    local_get(empty_list);
-                    i32_store(12);
-                    local_get(s);
-                });
-                self.scratch.free_i32(empty_list);
-                self.scratch.free_i32(s);
-            }
-            "json" => {
-                // http.json(status: Int, body: String) → Response (with Content-Type header)
-                let s = self.scratch.alloc_i32();
-                let hdr_list = self.scratch.alloc_i32();
-                let tuple_ptr = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(s);
-                    local_get(s);
-                });
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { i64_store(0); local_get(s); });
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { i32_store(8); });
-                // Build headers list with Content-Type: application/json
-                let ct_key = self.emitter.intern_string("Content-Type");
-                let ct_val = self.emitter.intern_string("application/json");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(tuple_ptr);
-                    local_get(tuple_ptr); i32_const(ct_key as i32); i32_store(0);
-                    local_get(tuple_ptr); i32_const(ct_val as i32); i32_store(4);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(hdr_list);
-                    local_get(hdr_list); i32_const(1); i32_store(0);
-                    local_get(hdr_list); local_get(tuple_ptr); i32_store(4);
-                    local_get(s); local_get(hdr_list); i32_store(12);
-                    local_get(s);
-                });
-                self.scratch.free_i32(tuple_ptr);
-                self.scratch.free_i32(hdr_list);
-                self.scratch.free_i32(s);
-            }
-            "status" if args.len() == 1 => {
-                // http.status(resp) → Int (getter)
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { i64_load(0); });
-            }
-            "status" if args.len() == 2 => {
-                // http.status(resp, new_status) → Response (setter)
-                let resp = self.scratch.alloc_i32();
-                let new_resp = self.scratch.alloc_i32();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(resp); });
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(new_resp);
-                    local_get(new_resp);
-                });
-                self.emit_expr(&args[1]); // new status: i64
-                wasm!(self.func, {
-                    i64_store(0);
-                    local_get(new_resp); local_get(resp); i32_load(8); i32_store(8);
-                    local_get(new_resp); local_get(resp); i32_load(12); i32_store(12);
-                    local_get(new_resp);
-                });
-                self.scratch.free_i32(new_resp);
-                self.scratch.free_i32(resp);
-            }
-            "body" => {
-                // http.body(resp) → String
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { i32_load(8); });
-            }
-            "get_header" => {
-                // http.get_header(resp, key) → Option[String]
-                let resp = self.scratch.alloc_i32();
-                let key = self.scratch.alloc_i32();
-                let hdrs = self.scratch.alloc_i32();
-                let len = self.scratch.alloc_i32();
-                let i = self.scratch.alloc_i32();
-                let pair_ptr = self.scratch.alloc_i32();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(resp); });
-                self.emit_expr(&args[1]);
-                wasm!(self.func, {
-                    local_set(key);
-                    local_get(resp); i32_load(12); local_set(hdrs);
-                    local_get(hdrs); i32_load(0); local_set(len);
-                    i32_const(0); local_set(i);
-                    block_empty; loop_empty;
-                      local_get(i); local_get(len); i32_ge_u; br_if(1);
-                      local_get(hdrs); i32_const(4); i32_add;
-                      local_get(i); i32_const(4); i32_mul; i32_add;
-                      i32_load(0); local_set(pair_ptr);
-                      local_get(pair_ptr); i32_load(0);
-                      local_get(key);
-                      call(self.emitter.rt.string.eq);
-                      if_empty;
-                        local_get(pair_ptr); i32_load(4); return_;
-                      end;
-                      local_get(i); i32_const(1); i32_add; local_set(i);
-                      br(0);
-                    end; end;
-                    i32_const(0); // none
-                });
-                self.scratch.free_i32(pair_ptr);
-                self.scratch.free_i32(i);
-                self.scratch.free_i32(len);
-                self.scratch.free_i32(hdrs);
-                self.scratch.free_i32(key);
-                self.scratch.free_i32(resp);
-            }
-            "set_header" => {
-                // http.set_header(resp, key, value) → Response (new response with header added/replaced)
-                let resp = self.scratch.alloc_i32();
-                let new_resp = self.scratch.alloc_i32();
-                let old_hdrs = self.scratch.alloc_i32();
-                let new_hdrs = self.scratch.alloc_i32();
-                let tuple_ptr = self.scratch.alloc_i32();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(resp); });
-                // Build new tuple (key, value)
-                let key_scratch = self.scratch.alloc_i32();
-                let val_scratch = self.scratch.alloc_i32();
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(key_scratch); });
-                self.emit_expr(&args[2]);
-                wasm!(self.func, {
-                    local_set(val_scratch);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(tuple_ptr);
-                    local_get(tuple_ptr); local_get(key_scratch); i32_store(0);
-                    local_get(tuple_ptr); local_get(val_scratch); i32_store(4);
-                    // Copy old headers + append new
-                    local_get(resp); i32_load(12); local_set(old_hdrs);
-                    local_get(old_hdrs); i32_load(0); i32_const(1); i32_add;
-                    local_set(val_scratch); // reuse as new_len
-                    i32_const(4); local_get(val_scratch); i32_const(4); i32_mul; i32_add;
-                    call(self.emitter.rt.alloc); local_set(new_hdrs);
-                    local_get(new_hdrs); local_get(val_scratch); i32_store(0);
-                    // Copy old header ptrs
-                    i32_const(0); local_set(key_scratch); // reuse as i
-                    block_empty; loop_empty;
-                      local_get(key_scratch); local_get(old_hdrs); i32_load(0); i32_ge_u; br_if(1);
-                      local_get(new_hdrs); i32_const(4); i32_add;
-                      local_get(key_scratch); i32_const(4); i32_mul; i32_add;
-                      local_get(old_hdrs); i32_const(4); i32_add;
-                      local_get(key_scratch); i32_const(4); i32_mul; i32_add;
-                      i32_load(0); i32_store(0);
-                      local_get(key_scratch); i32_const(1); i32_add; local_set(key_scratch);
-                      br(0);
-                    end; end;
-                    // Append new tuple at end
-                    local_get(new_hdrs); i32_const(4); i32_add;
-                    local_get(old_hdrs); i32_load(0); i32_const(4); i32_mul; i32_add;
-                    local_get(tuple_ptr); i32_store(0);
-                    // Build new response
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(new_resp);
-                    local_get(new_resp); local_get(resp); i64_load(0); i64_store(0);
-                    local_get(new_resp); local_get(resp); i32_load(8); i32_store(8);
-                    local_get(new_resp); local_get(new_hdrs); i32_store(12);
-                    local_get(new_resp);
-                });
-                self.scratch.free_i32(val_scratch);
-                self.scratch.free_i32(key_scratch);
-                self.scratch.free_i32(tuple_ptr);
-                self.scratch.free_i32(new_hdrs);
-                self.scratch.free_i32(old_hdrs);
-                self.scratch.free_i32(new_resp);
-                self.scratch.free_i32(resp);
-            }
-            "set_cookie" => {
-                // http.set_cookie(resp, name, value) = set_header(resp, "Set-Cookie", "name=value")
-                let resp = self.scratch.alloc_i32();
-                let cookie_val = self.scratch.alloc_i32();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(resp); });
-                self.emit_expr(&args[1]);
-                let eq_str = self.emitter.intern_string("=");
-                wasm!(self.func, { i32_const(eq_str as i32); call(self.emitter.rt.concat_str); });
-                self.emit_expr(&args[2]);
-                wasm!(self.func, { call(self.emitter.rt.concat_str); local_set(cookie_val); });
-                let cookie_key = self.emitter.intern_string("Set-Cookie");
-                let tuple_ptr = self.scratch.alloc_i32();
-                let new_hdrs = self.scratch.alloc_i32();
-                let old_hdrs = self.scratch.alloc_i32();
-                let new_resp = self.scratch.alloc_i32();
-                let new_len = self.scratch.alloc_i32();
-                let ci = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(tuple_ptr);
-                    local_get(tuple_ptr); i32_const(cookie_key as i32); i32_store(0);
-                    local_get(tuple_ptr); local_get(cookie_val); i32_store(4);
-                    local_get(resp); i32_load(12); local_set(old_hdrs);
-                    local_get(old_hdrs); i32_load(0); i32_const(1); i32_add; local_set(new_len);
-                    i32_const(4); local_get(new_len); i32_const(4); i32_mul; i32_add;
-                    call(self.emitter.rt.alloc); local_set(new_hdrs);
-                    local_get(new_hdrs); local_get(new_len); i32_store(0);
-                    i32_const(0); local_set(ci);
-                    block_empty; loop_empty;
-                      local_get(ci); local_get(old_hdrs); i32_load(0); i32_ge_u; br_if(1);
-                      local_get(new_hdrs); i32_const(4); i32_add; local_get(ci); i32_const(4); i32_mul; i32_add;
-                      local_get(old_hdrs); i32_const(4); i32_add; local_get(ci); i32_const(4); i32_mul; i32_add;
-                      i32_load(0); i32_store(0);
-                      local_get(ci); i32_const(1); i32_add; local_set(ci);
-                      br(0);
-                    end; end;
-                    local_get(new_hdrs); i32_const(4); i32_add;
-                    local_get(old_hdrs); i32_load(0); i32_const(4); i32_mul; i32_add;
-                    local_get(tuple_ptr); i32_store(0);
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(new_resp);
-                    local_get(new_resp); local_get(resp); i64_load(0); i64_store(0);
-                    local_get(new_resp); local_get(resp); i32_load(8); i32_store(8);
-                    local_get(new_resp); local_get(new_hdrs); i32_store(12);
-                    local_get(new_resp);
-                });
-                self.scratch.free_i32(ci);
-                self.scratch.free_i32(new_resp);
-                self.scratch.free_i32(old_hdrs);
-                self.scratch.free_i32(new_hdrs);
-                self.scratch.free_i32(new_len);
-                self.scratch.free_i32(tuple_ptr);
-                self.scratch.free_i32(cookie_val);
-                self.scratch.free_i32(resp);
-            }
-            "with_headers" => {
-                // http.with_headers(status, body, headers_map) → Response
-                // Map[String,String] layout: [len:i32][key0:i32][val0:i32][key1:i32][val1:i32]...
-                // Each entry is key_size=4 + val_size=4 = 8 bytes
-                let s = self.scratch.alloc_i32();
-                let hdr_map = self.scratch.alloc_i32();
-                let hdr_list = self.scratch.alloc_i32();
-                let len = self.scratch.alloc_i32();
-                let i = self.scratch.alloc_i32();
-                let tuple = self.scratch.alloc_i32();
-                wasm!(self.func, { i32_const(16); call(self.emitter.rt.alloc); local_set(s); local_get(s); });
-                self.emit_expr(&args[0]); // status
-                wasm!(self.func, { i64_store(0); local_get(s); });
-                self.emit_expr(&args[1]); // body
-                wasm!(self.func, { i32_store(8); });
-                self.emit_expr(&args[2]); // headers map
-                wasm!(self.func, {
-                    local_set(hdr_map);
-                    local_get(hdr_map); i32_load(0); local_set(len);
-                    // Build headers list from map entries
-                    i32_const(4); local_get(len); i32_const(4); i32_mul; i32_add;
-                    call(self.emitter.rt.alloc); local_set(hdr_list);
-                    local_get(hdr_list); local_get(len); i32_store(0);
-                    i32_const(0); local_set(i);
-                    block_empty; loop_empty;
-                      local_get(i); local_get(len); i32_ge_u; br_if(1);
-                      // Build tuple (key, val) from map entry
-                      i32_const(8); call(self.emitter.rt.alloc); local_set(tuple);
-                      local_get(tuple);
-                      local_get(hdr_map); i32_const(4); i32_add;
-                      local_get(i); i32_const(8); i32_mul; i32_add;
-                      i32_load(0); i32_store(0); // key
-                      local_get(tuple);
-                      local_get(hdr_map); i32_const(4); i32_add;
-                      local_get(i); i32_const(8); i32_mul; i32_add;
-                      i32_load(4); i32_store(4); // val
-                      local_get(hdr_list); i32_const(4); i32_add;
-                      local_get(i); i32_const(4); i32_mul; i32_add;
-                      local_get(tuple); i32_store(0);
-                      local_get(i); i32_const(1); i32_add; local_set(i);
-                      br(0);
-                    end; end;
-                    local_get(s); local_get(hdr_list); i32_store(12);
-                    local_get(s);
-                });
-                self.scratch.free_i32(tuple);
-                self.scratch.free_i32(i);
-                self.scratch.free_i32(len);
-                self.scratch.free_i32(hdr_list);
-                self.scratch.free_i32(hdr_map);
-                self.scratch.free_i32(s);
-            }
-            "not_found" => {
-                let s = self.scratch.alloc_i32();
-                wasm!(self.func, { i32_const(16); call(self.emitter.rt.alloc); local_set(s); local_get(s); i64_const(404); i64_store(0); local_get(s); });
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { i32_store(8); });
-                let empty = self.scratch.alloc_i32();
-                wasm!(self.func, { i32_const(4); call(self.emitter.rt.alloc); local_set(empty); local_get(empty); i32_const(0); i32_store(0); local_get(s); local_get(empty); i32_store(12); local_get(s); });
-                self.scratch.free_i32(empty);
-                self.scratch.free_i32(s);
-            }
-            "redirect" => {
-                let s = self.scratch.alloc_i32();
-                let empty_body = self.emitter.intern_string("");
-                wasm!(self.func, { i32_const(16); call(self.emitter.rt.alloc); local_set(s); local_get(s); i64_const(302); i64_store(0); local_get(s); i32_const(empty_body as i32); i32_store(8); });
-                let loc_key = self.emitter.intern_string("Location");
-                let tuple = self.scratch.alloc_i32();
-                let hdrs = self.scratch.alloc_i32();
-                wasm!(self.func, { i32_const(8); call(self.emitter.rt.alloc); local_set(tuple); local_get(tuple); i32_const(loc_key as i32); i32_store(0); local_get(tuple); });
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { i32_store(4); i32_const(8); call(self.emitter.rt.alloc); local_set(hdrs); local_get(hdrs); i32_const(1); i32_store(0); local_get(hdrs); local_get(tuple); i32_store(4); local_get(s); local_get(hdrs); i32_store(12); local_get(s); });
-                self.scratch.free_i32(hdrs);
-                self.scratch.free_i32(tuple);
-                self.scratch.free_i32(s);
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
-
-    /// fs module: read_text, write, exists
-    fn emit_fs_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "read_text" => {
-                // fs.read_text(path: String) -> Result[String, String]
-                // 1. Evaluate path arg (Almide String ptr: [len:i32][data:u8...])
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let file_size = self.scratch.alloc_i32();
-                let data_buf = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nread_ptr = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let str_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                // Allocate fd_out (4 bytes) via bump allocator
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // path_open(resolved_fd, dirflags=0, path_ptr, path_len, oflags=0,
-                //           rights=fd_read|fd_seek (2|4=6), inheriting=0, fdflags=0, fd_out_ptr)
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    i32_const(0);
-                    local_get(path_ptr);
-                    local_get(path_len);
-                    i32_const(0);
-                    i64_const(6);
-                    i64_const(0);
-                    i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                // If errno != 0, return err("file not found")
-                wasm!(self.func, {
-                    local_get(errno);
-                    i32_const(0);
-                    i32_ne;
-                    if_i32;
-                });
-                // Build err result
-                let err_msg = self.emitter.intern_string("file not found");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                // Load opened fd
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                });
-
-                // fd_filestat_get(fd, stat_buf) — stat_buf needs 64 bytes (allocator guarantees 8-byte alignment)
-                wasm!(self.func, {
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    local_get(opened_fd);
-                    local_get(stat_buf);
-                    call(self.emitter.rt.fd_filestat_get);
-                    drop;
-                });
-
-                // file_size = i32(stat_buf[32..40]) — file size is at offset 32 as i64, take lower 32 bits
-                wasm!(self.func, {
-                    local_get(stat_buf); i32_const(32); i32_add; i32_load(0); local_set(file_size);
-                });
-
-                // Allocate buffer for file data
-                wasm!(self.func, {
-                    local_get(file_size); call(self.emitter.rt.alloc); local_set(data_buf);
-                });
-
-                // Build iov struct: [buf_ptr:i32, buf_len:i32]
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    local_get(iov_ptr); local_get(data_buf); i32_store(0);
-                    local_get(iov_ptr); local_get(file_size); i32_store(4);
-                });
-
-                // nread_ptr
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                });
-
-                // fd_read(fd, iov_ptr, 1, nread_ptr)
-                wasm!(self.func, {
-                    local_get(opened_fd);
-                    local_get(iov_ptr);
-                    i32_const(1);
-                    local_get(nread_ptr);
-                    call(self.emitter.rt.fd_read);
-                    drop;
-                });
-
-                // fd_close(fd)
-                wasm!(self.func, {
-                    local_get(opened_fd);
-                    call(self.emitter.rt.fd_close);
-                    drop;
-                });
-
-                // Build Almide String: [len:i32][data:u8...]
-                // Use nread as actual length (may be <= file_size)
-                wasm!(self.func, {
-                    local_get(nread_ptr); i32_load(0); local_set(file_size);
-                    local_get(file_size); i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(str_ptr);
-                    local_get(str_ptr); local_get(file_size); i32_store(0);
-                });
-
-                // Copy data_buf[0..file_size] to str_ptr+4
-                // Byte-by-byte copy loop
-                let counter = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(0); local_set(counter);
-                    block_empty; loop_empty;
-                    local_get(counter); local_get(file_size); i32_ge_u; br_if(1);
-                    local_get(str_ptr); i32_const(4); i32_add; local_get(counter); i32_add;
-                    local_get(data_buf); local_get(counter); i32_add;
-                    i32_load8_u(0);
-                    i32_store8(0);
-                    local_get(counter); i32_const(1); i32_add; local_set(counter);
-                    br(0);
-                    end; end;
-                });
-                self.scratch.free_i32(counter);
-
-                // Build ok result: [tag=0:i32][str_ptr:i32]
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); local_get(str_ptr); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(str_ptr);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(data_buf);
-                self.scratch.free_i32(file_size);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "write" => {
-                // fs.write(path: String, content: String) -> Result[Unit, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let content_str = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nwritten_ptr = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                // Evaluate path
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                // Evaluate content
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(content_str); });
-
-                // Allocate fd_out
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // path_open(resolved_fd, dirflags=0, path_ptr, path_len,
-                //           oflags=O_CREAT|O_TRUNC(=9),
-                //           rights=fd_write(=64), inheriting=0, fdflags=0, fd_out_ptr)
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    i32_const(0);
-                    local_get(path_ptr);
-                    local_get(path_len);
-                    i32_const(9);
-                    i64_const(64);
-                    i64_const(0);
-                    i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                // If errno != 0, return err
-                wasm!(self.func, {
-                    local_get(errno);
-                    i32_const(0);
-                    i32_ne;
-                    if_i32;
-                });
-                let err_msg = self.emitter.intern_string("failed to open file for writing");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                // Load opened fd
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                });
-
-                // Build iov: [content_ptr+4, content_len]
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    local_get(iov_ptr); local_get(content_str); i32_const(4); i32_add; i32_store(0);
-                    local_get(iov_ptr); local_get(content_str); i32_load(0); i32_store(4);
-                });
-
-                // nwritten_ptr
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nwritten_ptr);
-                });
-
-                // fd_write(fd, iov_ptr, 1, nwritten_ptr)
-                wasm!(self.func, {
-                    local_get(opened_fd);
-                    local_get(iov_ptr);
-                    i32_const(1);
-                    local_get(nwritten_ptr);
-                    call(self.emitter.rt.fd_write);
-                    drop;
-                });
-
-                // fd_close(fd)
-                wasm!(self.func, {
-                    local_get(opened_fd);
-                    call(self.emitter.rt.fd_close);
-                    drop;
-                });
-
-                // Build ok(unit) result: [tag=0:i32][0:i32]
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(nwritten_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(content_str);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "exists" => {
-                // fs.exists(path: String) -> Bool
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                // Allocate 64-byte stat buffer (allocator guarantees 8-byte alignment)
-                wasm!(self.func, {
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                });
-
-                // path_filestat_get(resolved_fd, flags=0, path_ptr, path_len, stat_buf)
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    i32_const(0);
-                    local_get(path_ptr);
-                    local_get(path_len);
-                    local_get(stat_buf);
-                    call(self.emitter.rt.path_filestat_get);
-                    // errno == 0 → true (1), else false (0)
-                    i32_eqz;
-                });
-
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "read_bytes" => {
-                // fs.read_bytes(path) -> Result[List[Int], String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let file_size = self.scratch.alloc_i32();
-                let data_buf = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nread_ptr = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let list_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-                let counter = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // path_open for reading
-                wasm!(self.func, {
-                    local_get(resolved_fd); i32_const(0);
-                    local_get(path_ptr); local_get(path_len);
-                    i32_const(0); i64_const(6); i64_const(0); i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                });
-                let err_msg = self.emitter.intern_string("file not found");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                // stat for file size
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    local_get(opened_fd); local_get(stat_buf);
-                    call(self.emitter.rt.fd_filestat_get); drop;
-                    local_get(stat_buf); i32_const(32); i32_add; i32_load(0); local_set(file_size);
-                });
-
-                // Read raw bytes
-                wasm!(self.func, {
-                    local_get(file_size); call(self.emitter.rt.alloc); local_set(data_buf);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    local_get(iov_ptr); local_get(data_buf); i32_store(0);
-                    local_get(iov_ptr); local_get(file_size); i32_store(4);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                    local_get(opened_fd); local_get(iov_ptr); i32_const(1); local_get(nread_ptr);
-                    call(self.emitter.rt.fd_read); drop;
-                    local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-                    local_get(nread_ptr); i32_load(0); local_set(file_size);
-                });
-
-                // Build List[Int]: [count:i32][i64 * count]
-                wasm!(self.func, {
-                    local_get(file_size); i32_const(8); i32_mul; i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(list_ptr);
-                    local_get(list_ptr); local_get(file_size); i32_store(0);
-                });
-
-                // Copy each byte as i64
-                wasm!(self.func, {
-                    i32_const(0); local_set(counter);
-                    block_empty; loop_empty;
-                    local_get(counter); local_get(file_size); i32_ge_u; br_if(1);
-                    local_get(list_ptr); i32_const(4); i32_add;
-                    local_get(counter); i32_const(8); i32_mul; i32_add;
-                    local_get(data_buf); local_get(counter); i32_add; i32_load8_u(0);
-                    i64_extend_i32_u;
-                    i64_store(0);
-                    local_get(counter); i32_const(1); i32_add; local_set(counter);
-                    br(0);
-                    end; end;
-                });
-
-                // ok(list_ptr)
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); local_get(list_ptr); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(counter);
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(list_ptr);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(data_buf);
-                self.scratch.free_i32(file_size);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "write_bytes" => {
-                // fs.write_bytes(path, bytes: List[Int]) -> Result[Unit, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let list_ptr = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nwritten_ptr = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-                let byte_buf = self.scratch.alloc_i32();
-                let count = self.scratch.alloc_i32();
-                let counter = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(list_ptr); });
-
-                // Convert List[Int] (i64 elements) to byte buffer
-                wasm!(self.func, {
-                    local_get(list_ptr); i32_load(0); local_set(count);
-                    local_get(count); call(self.emitter.rt.alloc); local_set(byte_buf);
-                    i32_const(0); local_set(counter);
-                    block_empty; loop_empty;
-                    local_get(counter); local_get(count); i32_ge_u; br_if(1);
-                    local_get(byte_buf); local_get(counter); i32_add;
-                    local_get(list_ptr); i32_const(4); i32_add;
-                    local_get(counter); i32_const(8); i32_mul; i32_add;
-                    i64_load(0); i32_wrap_i64;
-                    i32_store8(0);
-                    local_get(counter); i32_const(1); i32_add; local_set(counter);
-                    br(0);
-                    end; end;
-                });
-
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // path_open for writing (O_CREAT|O_TRUNC=9)
-                wasm!(self.func, {
-                    local_get(resolved_fd); i32_const(0);
-                    local_get(path_ptr); local_get(path_len);
-                    i32_const(9); i64_const(64); i64_const(0); i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                });
-                let err_msg = self.emitter.intern_string("failed to open file for writing");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    local_get(iov_ptr); local_get(byte_buf); i32_store(0);
-                    local_get(iov_ptr); local_get(count); i32_store(4);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nwritten_ptr);
-                    local_get(opened_fd); local_get(iov_ptr); i32_const(1); local_get(nwritten_ptr);
-                    call(self.emitter.rt.fd_write); drop;
-                    local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-                });
-
-                // ok(unit)
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(counter);
-                self.scratch.free_i32(count);
-                self.scratch.free_i32(byte_buf);
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(nwritten_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(list_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "append" => {
-                // fs.append(path, content) -> Result[Unit, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let content_str = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nwritten_ptr = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(content_str); });
-
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // path_open: oflags=O_CREAT(1), rights=fd_write(64), fdflags=APPEND(1)
-                wasm!(self.func, {
-                    local_get(resolved_fd); i32_const(0);
-                    local_get(path_ptr); local_get(path_len);
-                    i32_const(1);
-                    i64_const(64); i64_const(0);
-                    i32_const(1);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                });
-                let err_msg = self.emitter.intern_string("failed to open file for appending");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    local_get(iov_ptr); local_get(content_str); i32_const(4); i32_add; i32_store(0);
-                    local_get(iov_ptr); local_get(content_str); i32_load(0); i32_store(4);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nwritten_ptr);
-                    local_get(opened_fd); local_get(iov_ptr); i32_const(1); local_get(nwritten_ptr);
-                    call(self.emitter.rt.fd_write); drop;
-                    local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-                });
-
-                // ok(unit)
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(nwritten_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(content_str);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "mkdir_p" => {
-                // fs.mkdir_p(path) -> Result[Unit, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                // Iterative mkdir_p: create each prefix segment
-                let seg_end = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(0); local_set(seg_end);
-                    block_empty; loop_empty;
-                    local_get(seg_end); local_get(path_len); i32_ge_u; br_if(1);
-                    // Advance seg_end past current char
-                    local_get(seg_end); i32_const(1); i32_add; local_set(seg_end);
-                    // Skip to next '/' or end of path
-                    block_empty; loop_empty;
-                    local_get(seg_end); local_get(path_len); i32_ge_u; br_if(1);
-                    local_get(path_ptr); local_get(seg_end); i32_add; i32_load8_u(0);
-                    i32_const(47); i32_eq; br_if(1);
-                    local_get(seg_end); i32_const(1); i32_add; local_set(seg_end);
-                    br(0);
-                    end; end;
-                    // Try creating directory for path[0..seg_end]
-                    local_get(resolved_fd);
-                    local_get(path_ptr);
-                    local_get(seg_end);
-                    call(self.emitter.rt.path_create_directory);
-                    drop;
-                    br(0);
-                    end; end;
-                });
-                self.scratch.free_i32(seg_end);
-
-                // Final attempt: create the full path and check error
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    local_get(path_ptr);
-                    local_get(path_len);
-                    call(self.emitter.rt.path_create_directory);
-                    local_set(errno);
-                });
-
-                // errno==0 or errno==20 (EEXIST) -> ok
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    local_get(errno); i32_const(20); i32_eq;
-                    i32_or;
-                    if_i32;
-                });
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("failed to create directory");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "read_lines" => {
-                // fs.read_lines(path) -> Result[List[String], String]
-                // Call read_text internally, then split by '\n' using string.lines
-                self.emit_fs_call_inner_read_text(args);
-                let res = self.scratch.alloc_i32();
-                let tag = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    local_set(res);
-                    local_get(res); i32_load(0); local_set(tag);
-                    local_get(tag); i32_eqz;
-                    if_i32;
-                });
-                // ok path: split the string by '\n'
-                let text_ptr = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    local_get(res); i32_const(4); i32_add; i32_load(0); local_set(text_ptr);
-                    local_get(text_ptr);
-                    call(self.emitter.rt.string.lines);
-                });
-                let result_ptr = self.scratch.alloc_i32();
-                let list_val = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    local_set(list_val);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); local_get(list_val); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                    local_get(res);
-                    end;
-                });
-
-                self.scratch.free_i32(list_val);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(text_ptr);
-                self.scratch.free_i32(tag);
-                self.scratch.free_i32(res);
-            }
-            "list_dir" => {
-                // fs.list_dir(path) -> Result[List[String], String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-                let dir_buf = self.scratch.alloc_i32();
-                let bufused_ptr = self.scratch.alloc_i32();
-                let bufused = self.scratch.alloc_i32();
-                let offset = self.scratch.alloc_i32();
-                let list_ptr = self.scratch.alloc_i32();
-                let list_count = self.scratch.alloc_i32();
-                let entry_name_len = self.scratch.alloc_i32();
-                let str_ptr = self.scratch.alloc_i32();
-                let counter = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // path_open for directory: dirflags=1(symlink follow), oflags=O_DIRECTORY(2)
-                // rights = fd_readdir(0x4000)
-                wasm!(self.func, {
-                    local_get(resolved_fd); i32_const(1);
-                    local_get(path_ptr); local_get(path_len);
-                    i32_const(2);
-                    i64_const(0x4000);
-                    i64_const(0);
-                    i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                });
-                let err_msg = self.emitter.intern_string("failed to open directory");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                });
-
-                // Allocate readdir buffer (4KB) and bufused output
-                wasm!(self.func, {
-                    i32_const(4096); call(self.emitter.rt.alloc); local_set(dir_buf);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(bufused_ptr);
-                });
-
-                // fd_readdir(fd, buf, buf_len, cookie=0, bufused_ptr)
-                wasm!(self.func, {
-                    local_get(opened_fd);
-                    local_get(dir_buf);
-                    i32_const(4096);
-                    i64_const(0);
-                    local_get(bufused_ptr);
-                    call(self.emitter.rt.fd_readdir);
-                    drop;
-                    local_get(bufused_ptr); i32_load(0); local_set(bufused);
-                    local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-                });
-
-                // First pass: count entries (skipping "." and "..")
-                // WASI dirent: d_next(8) + d_ino(8) + d_namlen(4) + d_type(4) = 24 bytes header
-                let skip = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(0); local_set(offset);
-                    i32_const(0); local_set(list_count);
-                    block_empty; loop_empty;
-                    local_get(offset); i32_const(24); i32_add;
-                    local_get(bufused); i32_gt_u; br_if(1);
-                    local_get(dir_buf); local_get(offset); i32_add; i32_const(16); i32_add;
-                    i32_load(0); local_set(entry_name_len);
-
-                    // skip = (namlen==1 && name[0]=='.') || (namlen==2 && name[0]=='.' && name[1]=='.')
-                    i32_const(0); local_set(skip);
-                    // Check "."
-                    local_get(entry_name_len); i32_const(1); i32_eq;
-                    if_empty;
-                      local_get(dir_buf); local_get(offset); i32_add; i32_const(24); i32_add;
-                      i32_load8_u(0); i32_const(46); i32_eq;
-                      if_empty; i32_const(1); local_set(skip); end;
-                    end;
-                    // Check ".."
-                    local_get(entry_name_len); i32_const(2); i32_eq;
-                    if_empty;
-                      local_get(dir_buf); local_get(offset); i32_add; i32_const(24); i32_add;
-                      i32_load8_u(0); i32_const(46); i32_eq;
-                      local_get(dir_buf); local_get(offset); i32_add; i32_const(25); i32_add;
-                      i32_load8_u(0); i32_const(46); i32_eq;
-                      i32_and;
-                      if_empty; i32_const(1); local_set(skip); end;
-                    end;
-                    // Count if not skipped
-                    local_get(skip); i32_eqz;
-                    if_empty;
-                      local_get(list_count); i32_const(1); i32_add; local_set(list_count);
-                    end;
-
-                    // Advance offset
-                    local_get(offset); i32_const(24); i32_add; local_get(entry_name_len); i32_add;
-                    local_set(offset);
-                    br(0);
-                    end; end;
-                });
-
-                // Allocate List[String]: [count:i32][ptr:i32 * count]
-                wasm!(self.func, {
-                    local_get(list_count); i32_const(4); i32_mul; i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(list_ptr);
-                    local_get(list_ptr); local_get(list_count); i32_store(0);
-                });
-
-                // Second pass: build string entries (same skip logic as counting pass)
-                let copy_i = self.scratch.alloc_i32();
-                wasm!(self.func, {
-                    i32_const(0); local_set(offset);
-                    i32_const(0); local_set(counter);
-                    block_empty; loop_empty;
-                    local_get(offset); i32_const(24); i32_add;
-                    local_get(bufused); i32_gt_u; br_if(1);
-                    local_get(dir_buf); local_get(offset); i32_add; i32_const(16); i32_add;
-                    i32_load(0); local_set(entry_name_len);
-
-                    // skip = (namlen==1 && name[0]=='.') || (namlen==2 && name[0]=='.' && name[1]=='.')
-                    i32_const(0); local_set(skip);
-                    local_get(entry_name_len); i32_const(1); i32_eq;
-                    if_empty;
-                      local_get(dir_buf); local_get(offset); i32_add; i32_const(24); i32_add;
-                      i32_load8_u(0); i32_const(46); i32_eq;
-                      if_empty; i32_const(1); local_set(skip); end;
-                    end;
-                    local_get(entry_name_len); i32_const(2); i32_eq;
-                    if_empty;
-                      local_get(dir_buf); local_get(offset); i32_add; i32_const(24); i32_add;
-                      i32_load8_u(0); i32_const(46); i32_eq;
-                      local_get(dir_buf); local_get(offset); i32_add; i32_const(25); i32_add;
-                      i32_load8_u(0); i32_const(46); i32_eq;
-                      i32_and;
-                      if_empty; i32_const(1); local_set(skip); end;
-                    end;
-                    // Build entry if not skipped
-                    local_get(skip); i32_eqz;
-                    if_empty;
-                });
-                self.emit_fs_list_dir_build_entry(copy_i, entry_name_len, str_ptr, dir_buf, offset, list_ptr, counter);
-                wasm!(self.func, {
-                    end;
-
-                    // Advance offset
-                    local_get(offset); i32_const(24); i32_add; local_get(entry_name_len); i32_add;
-                    local_set(offset);
-                    br(0);
-                    end; end;
-                });
-                self.scratch.free_i32(copy_i);
-
-                // Update list count
-                wasm!(self.func, {
-                    local_get(list_ptr); local_get(counter); i32_store(0);
-                });
-
-                // ok(list_ptr)
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); local_get(list_ptr); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(skip);
-                self.scratch.free_i32(counter);
-                self.scratch.free_i32(str_ptr);
-                self.scratch.free_i32(entry_name_len);
-                self.scratch.free_i32(list_count);
-                self.scratch.free_i32(list_ptr);
-                self.scratch.free_i32(offset);
-                self.scratch.free_i32(bufused);
-                self.scratch.free_i32(bufused_ptr);
-                self.scratch.free_i32(dir_buf);
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "is_dir" => {
-                // fs.is_dir(path) -> Bool  (filetype 3 = directory)
-                self.emit_fs_filetype_check(args, 3);
-            }
-            "is_file" => {
-                // fs.is_file(path) -> Bool  (filetype 4 = regular file)
-                self.emit_fs_filetype_check(args, 4);
-            }
-            "is_symlink" => {
-                // fs.is_symlink(path) -> Bool  (filetype 7 = symbolic_link)
-                // Use flags=0 (do NOT follow symlinks)
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    // flags=0: do NOT follow symlinks
-                    local_get(resolved_fd); i32_const(0);
-                    local_get(path_ptr); local_get(path_len);
-                    local_get(stat_buf);
-                    call(self.emitter.rt.path_filestat_get);
-                    local_set(errno);
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                      i32_const(0);
-                    else_;
-                      local_get(stat_buf); i32_const(16); i32_add; i32_load8_u(0);
-                      i32_const(7); i32_eq;
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "copy" => {
-                // fs.copy(src, dst) -> Result[Unit, String]
-                // Read source file bytes, write to destination
-                let src_str = self.scratch.alloc_i32();
-                let src_ptr = self.scratch.alloc_i32();
-                let src_len = self.scratch.alloc_i32();
-                let src_resolved_fd = self.scratch.alloc_i32();
-                let dst_str = self.scratch.alloc_i32();
-                let dst_ptr = self.scratch.alloc_i32();
-                let dst_len = self.scratch.alloc_i32();
-                let dst_resolved_fd = self.scratch.alloc_i32();
-                let fd_out_ptr = self.scratch.alloc_i32();
-                let opened_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let file_size = self.scratch.alloc_i32();
-                let data_buf = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nrw_ptr = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(src_str); });
-                self.emit_fs_resolve_path(src_str, src_ptr, src_len, src_resolved_fd);
-
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(dst_str); });
-                self.emit_fs_resolve_path(dst_str, dst_ptr, dst_len, dst_resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-                });
-
-                // Open source for reading
-                wasm!(self.func, {
-                    local_get(src_resolved_fd); i32_const(0);
-                    local_get(src_ptr); local_get(src_len);
-                    i32_const(0); i64_const(6); i64_const(0); i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                });
-                let err_msg = self.emitter.intern_string("failed to open source file");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                // Read source content
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    local_get(opened_fd); local_get(stat_buf);
-                    call(self.emitter.rt.fd_filestat_get); drop;
-                    local_get(stat_buf); i32_const(32); i32_add; i32_load(0); local_set(file_size);
-                    local_get(file_size); call(self.emitter.rt.alloc); local_set(data_buf);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    local_get(iov_ptr); local_get(data_buf); i32_store(0);
-                    local_get(iov_ptr); local_get(file_size); i32_store(4);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nrw_ptr);
-                    local_get(opened_fd); local_get(iov_ptr); i32_const(1); local_get(nrw_ptr);
-                    call(self.emitter.rt.fd_read); drop;
-                    local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-                    local_get(nrw_ptr); i32_load(0); local_set(file_size);
-                });
-
-                // Open dst for writing
-                wasm!(self.func, {
-                    local_get(dst_resolved_fd); i32_const(0);
-                    local_get(dst_ptr); local_get(dst_len);
-                    i32_const(9); i64_const(64); i64_const(0); i32_const(0);
-                    local_get(fd_out_ptr);
-                    call(self.emitter.rt.path_open);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_const(0); i32_ne;
-                    if_i32;
-                });
-                let err_msg2 = self.emitter.intern_string("failed to open destination file");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg2 as i32); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                // Write data to dst
-                wasm!(self.func, {
-                    local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-                    local_get(iov_ptr); local_get(data_buf); i32_store(0);
-                    local_get(iov_ptr); local_get(file_size); i32_store(4);
-                    local_get(opened_fd); local_get(iov_ptr); i32_const(1); local_get(nrw_ptr);
-                    call(self.emitter.rt.fd_write); drop;
-                    local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-                });
-
-                // ok(unit) -- close nested if blocks
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(nrw_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(data_buf);
-                self.scratch.free_i32(file_size);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(opened_fd);
-                self.scratch.free_i32(fd_out_ptr);
-                self.scratch.free_i32(dst_resolved_fd);
-                self.scratch.free_i32(dst_len);
-                self.scratch.free_i32(dst_ptr);
-                self.scratch.free_i32(dst_str);
-                self.scratch.free_i32(src_resolved_fd);
-                self.scratch.free_i32(src_len);
-                self.scratch.free_i32(src_ptr);
-                self.scratch.free_i32(src_str);
-            }
-            "rename" => {
-                // fs.rename(src, dst) -> Result[Unit, String]
-                let src_str = self.scratch.alloc_i32();
-                let src_ptr = self.scratch.alloc_i32();
-                let src_len = self.scratch.alloc_i32();
-                let src_resolved_fd = self.scratch.alloc_i32();
-                let dst_str = self.scratch.alloc_i32();
-                let dst_ptr = self.scratch.alloc_i32();
-                let dst_len = self.scratch.alloc_i32();
-                let dst_resolved_fd = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(src_str); });
-                self.emit_fs_resolve_path(src_str, src_ptr, src_len, src_resolved_fd);
-
-                self.emit_expr(&args[1]);
-                wasm!(self.func, { local_set(dst_str); });
-                self.emit_fs_resolve_path(dst_str, dst_ptr, dst_len, dst_resolved_fd);
-
-                // path_rename(old_fd, old_path, old_len, new_fd, new_path, new_len)
-                wasm!(self.func, {
-                    local_get(src_resolved_fd);
-                    local_get(src_ptr); local_get(src_len);
-                    local_get(dst_resolved_fd);
-                    local_get(dst_ptr); local_get(dst_len);
-                    call(self.emitter.rt.path_rename);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("failed to rename");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(dst_resolved_fd);
-                self.scratch.free_i32(dst_len);
-                self.scratch.free_i32(dst_ptr);
-                self.scratch.free_i32(dst_str);
-                self.scratch.free_i32(src_resolved_fd);
-                self.scratch.free_i32(src_len);
-                self.scratch.free_i32(src_ptr);
-                self.scratch.free_i32(src_str);
-            }
-            "remove" => {
-                // fs.remove(path) -> Result[Unit, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    local_get(path_ptr); local_get(path_len);
-                    call(self.emitter.rt.path_unlink_file);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("failed to remove file");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "remove_all" => {
-                // fs.remove_all(path) -> Result[Unit, String]
-                // Try unlink (file), then rmdir (empty dir)
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                // Try path_unlink_file first
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    local_get(path_ptr); local_get(path_len);
-                    call(self.emitter.rt.path_unlink_file);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-
-                // Try path_remove_directory
-                wasm!(self.func, {
-                    local_get(resolved_fd);
-                    local_get(path_ptr); local_get(path_len);
-                    call(self.emitter.rt.path_remove_directory);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(0); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("failed to remove path");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "file_size" => {
-                // fs.file_size(path) -> Result[Int, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    local_get(resolved_fd); i32_const(1);
-                    local_get(path_ptr); local_get(path_len);
-                    local_get(stat_buf);
-                    call(self.emitter.rt.path_filestat_get);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-                // ok: file size at offset 32 as i64
-                // Result[Int, String] = [tag:i32][padding:i32][i64] = 16 bytes
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(8); i32_add;
-                    local_get(stat_buf); i32_const(32); i32_add; i64_load(0);
-                    i64_store(0);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("file not found");
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(8); i32_add;
-                    i32_const(err_msg as i32); i64_extend_i32_u; i64_store(0);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "modified_at" => {
-                // fs.modified_at(path) -> Result[Int, String]
-                // mtim at offset 40 (u64, nanoseconds) -> seconds
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    local_get(resolved_fd); i32_const(1);
-                    local_get(path_ptr); local_get(path_len);
-                    local_get(stat_buf);
-                    call(self.emitter.rt.path_filestat_get);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); i32_const(8); i32_add;
-                    local_get(stat_buf); i32_const(40); i32_add; i64_load(0);
-                    i64_const(1000000000); i64_div_u;
-                    i64_store(0);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("file not found");
-                wasm!(self.func, {
-                    i32_const(16); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(8); i32_add;
-                    i32_const(err_msg as i32); i64_extend_i32_u; i64_store(0);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "stat" => {
-                // fs.stat(path) -> Result[{size: Int, is_dir: Bool, is_file: Bool, modified: Int}, String]
-                let path_str = self.scratch.alloc_i32();
-                let path_ptr = self.scratch.alloc_i32();
-                let path_len = self.scratch.alloc_i32();
-                let resolved_fd = self.scratch.alloc_i32();
-                let stat_buf = self.scratch.alloc_i32();
-                let result_ptr = self.scratch.alloc_i32();
-                let rec_ptr = self.scratch.alloc_i32();
-                let errno = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, { local_set(path_str); });
-                self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-                wasm!(self.func, {
-                    i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-                    local_get(resolved_fd); i32_const(1);
-                    local_get(path_ptr); local_get(path_len);
-                    local_get(stat_buf);
-                    call(self.emitter.rt.path_filestat_get);
-                    local_set(errno);
-                });
-
-                wasm!(self.func, {
-                    local_get(errno); i32_eqz;
-                    if_i32;
-                });
-
-                // Record: [size:i64(8)][is_dir:i32(4)][is_file:i32(4)][modified:i64(8)] = 24 bytes
-                wasm!(self.func, {
-                    i32_const(24); call(self.emitter.rt.alloc); local_set(rec_ptr);
-                    // size at stat offset 32
-                    local_get(rec_ptr);
-                    local_get(stat_buf); i32_const(32); i32_add; i64_load(0);
-                    i64_store(0);
-                    // is_dir: filetype at offset 16 == 3
-                    local_get(rec_ptr); i32_const(8); i32_add;
-                    local_get(stat_buf); i32_const(16); i32_add; i32_load8_u(0);
-                    i32_const(3); i32_eq;
-                    i32_store(0);
-                    // is_file: filetype at offset 16 == 4
-                    local_get(rec_ptr); i32_const(12); i32_add;
-                    local_get(stat_buf); i32_const(16); i32_add; i32_load8_u(0);
-                    i32_const(4); i32_eq;
-                    i32_store(0);
-                    // modified: mtim at stat offset 40, nanoseconds -> seconds
-                    local_get(rec_ptr); i32_const(16); i32_add;
-                    local_get(stat_buf); i32_const(40); i32_add; i64_load(0);
-                    i64_const(1000000000); i64_div_u;
-                    i64_store(0);
-                });
-
-                // ok(rec_ptr)
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(0); i32_store(0);
-                    local_get(result_ptr); local_get(rec_ptr); i32_store(4);
-                    local_get(result_ptr);
-                    else_;
-                });
-                let err_msg = self.emitter.intern_string("file not found");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                    end;
-                });
-
-                self.scratch.free_i32(errno);
-                self.scratch.free_i32(rec_ptr);
-                self.scratch.free_i32(result_ptr);
-                self.scratch.free_i32(stat_buf);
-                self.scratch.free_i32(resolved_fd);
-                self.scratch.free_i32(path_len);
-                self.scratch.free_i32(path_ptr);
-                self.scratch.free_i32(path_str);
-            }
-            "walk" | "glob" | "create_temp_file" | "create_temp_dir" => {
-                // These require recursive dir traversal (walk), glob pattern matching (glob),
-                // or OS temp dir + random naming (create_temp_*) which are infeasible in pure WASI.
-                for arg in args { self.emit_expr(arg); if super::values::ty_to_valtype(&arg.ty).is_some() { wasm!(self.func, { drop; }); } }
-                let result_ptr = self.scratch.alloc_i32();
-                let err_msg = self.emitter.intern_string("not supported in WASM");
-                wasm!(self.func, {
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-                    local_get(result_ptr); i32_const(1); i32_store(0);
-                    local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-                    local_get(result_ptr);
-                });
-                self.scratch.free_i32(result_ptr);
-            }
-            "temp_dir" => {
-                // fs.temp_dir() -> String: return "/tmp"
-                let s = self.emitter.intern_string("/tmp");
-                wasm!(self.func, { i32_const(s as i32); });
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
-
-    /// Helper: resolve path using __resolve_path runtime function.
-    /// After calling, the scratch locals contain:
-    /// - resolved_fd: the preopened dir fd
-    /// - path_ptr: pointer to the relative path bytes
-    /// - path_len: length of the relative path
-    /// This replaces the old pattern of hardcoded fd=3 + strip leading '/'.
-    fn emit_fs_resolve_path(&mut self, path_str: u32, path_ptr: u32, path_len: u32, resolved_fd: u32) {
-        let resolve_result = self.scratch.alloc_i32();
-        wasm!(self.func, {
-            // Call __resolve_path(path_ptr, path_len) → result_ptr [fd, rel_ptr, rel_len]
-            local_get(path_str); i32_const(4); i32_add; // raw path bytes (skip string length prefix)
-            local_get(path_str); i32_load(0);            // path byte length
-            call(self.emitter.rt.resolve_path);
-            local_set(resolve_result);
-            // Unpack result
-            local_get(resolve_result); i32_load(0); local_set(resolved_fd);
-            local_get(resolve_result); i32_load(4); local_set(path_ptr);
-            local_get(resolve_result); i32_load(8); local_set(path_len);
-        });
-        self.scratch.free_i32(resolve_result);
-    }
-
-    /// Helper: check path filetype against expected value. Used by is_dir, is_file.
-    fn emit_fs_filetype_check(&mut self, args: &[IrExpr], expected_filetype: i32) {
-        let path_str = self.scratch.alloc_i32();
-        let path_ptr = self.scratch.alloc_i32();
-        let path_len = self.scratch.alloc_i32();
-        let resolved_fd = self.scratch.alloc_i32();
-        let stat_buf = self.scratch.alloc_i32();
-        let errno = self.scratch.alloc_i32();
-
-        self.emit_expr(&args[0]);
-        wasm!(self.func, { local_set(path_str); });
-        self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-        wasm!(self.func, {
-            i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-            // flags=1 (follow symlinks) for is_dir/is_file
-            local_get(resolved_fd); i32_const(1);
-            local_get(path_ptr); local_get(path_len);
-            local_get(stat_buf);
-            call(self.emitter.rt.path_filestat_get);
-            local_set(errno);
-            local_get(errno); i32_const(0); i32_ne;
-            if_i32;
-              i32_const(0);
-            else_;
-              // filetype at stat offset 16
-              local_get(stat_buf); i32_const(16); i32_add; i32_load8_u(0);
-              i32_const(expected_filetype);
-              i32_eq;
-            end;
-        });
-
-        self.scratch.free_i32(errno);
-        self.scratch.free_i32(stat_buf);
-        self.scratch.free_i32(resolved_fd);
-        self.scratch.free_i32(path_len);
-        self.scratch.free_i32(path_ptr);
-        self.scratch.free_i32(path_str);
-    }
-
-    /// Helper for list_dir: build a string entry from dirent name and store into list.
-    fn emit_fs_list_dir_build_entry(
-        &mut self,
-        copy_i: u32, entry_name_len: u32, str_ptr: u32,
-        dir_buf: u32, offset: u32, list_ptr: u32, counter: u32,
-    ) {
-        wasm!(self.func, {
-            local_get(entry_name_len); i32_const(4); i32_add;
-            call(self.emitter.rt.alloc); local_set(str_ptr);
-            local_get(str_ptr); local_get(entry_name_len); i32_store(0);
-            // Copy name bytes
-            i32_const(0); local_set(copy_i);
-            block_empty; loop_empty;
-            local_get(copy_i); local_get(entry_name_len); i32_ge_u; br_if(1);
-            local_get(str_ptr); i32_const(4); i32_add; local_get(copy_i); i32_add;
-            local_get(dir_buf); local_get(offset); i32_add; i32_const(24); i32_add;
-            local_get(copy_i); i32_add; i32_load8_u(0);
-            i32_store8(0);
-            local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-            br(0);
-            end; end;
-            // Store in list
-            local_get(list_ptr); i32_const(4); i32_add;
-            local_get(counter); i32_const(4); i32_mul; i32_add;
-            local_get(str_ptr); i32_store(0);
-            local_get(counter); i32_const(1); i32_add; local_set(counter);
-        });
-    }
-
-    /// Helper: emit read_text logic, leaving Result[String, String] on stack.
-    fn emit_fs_call_inner_read_text(&mut self, args: &[IrExpr]) {
-        let path_str = self.scratch.alloc_i32();
-        let path_ptr = self.scratch.alloc_i32();
-        let path_len = self.scratch.alloc_i32();
-        let resolved_fd = self.scratch.alloc_i32();
-        let fd_out_ptr = self.scratch.alloc_i32();
-        let opened_fd = self.scratch.alloc_i32();
-        let stat_buf = self.scratch.alloc_i32();
-        let file_size = self.scratch.alloc_i32();
-        let data_buf = self.scratch.alloc_i32();
-        let iov_ptr = self.scratch.alloc_i32();
-        let nread_ptr = self.scratch.alloc_i32();
-        let result_ptr = self.scratch.alloc_i32();
-        let str_ptr = self.scratch.alloc_i32();
-        let errno = self.scratch.alloc_i32();
-
-        self.emit_expr(&args[0]);
-        wasm!(self.func, { local_set(path_str); });
-        self.emit_fs_resolve_path(path_str, path_ptr, path_len, resolved_fd);
-
-        wasm!(self.func, {
-            i32_const(4); call(self.emitter.rt.alloc); local_set(fd_out_ptr);
-        });
-
-        wasm!(self.func, {
-            local_get(resolved_fd); i32_const(0);
-            local_get(path_ptr); local_get(path_len);
-            i32_const(0); i64_const(6); i64_const(0); i32_const(0);
-            local_get(fd_out_ptr);
-            call(self.emitter.rt.path_open);
-            local_set(errno);
-        });
-
-        wasm!(self.func, {
-            local_get(errno); i32_const(0); i32_ne;
-            if_i32;
-        });
-        let err_msg = self.emitter.intern_string("file not found");
-        wasm!(self.func, {
-            i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-            local_get(result_ptr); i32_const(1); i32_store(0);
-            local_get(result_ptr); i32_const(err_msg as i32); i32_store(4);
-            local_get(result_ptr);
-            else_;
-        });
-
-        wasm!(self.func, {
-            local_get(fd_out_ptr); i32_load(0); local_set(opened_fd);
-            i32_const(64); call(self.emitter.rt.alloc); local_set(stat_buf);
-            local_get(opened_fd); local_get(stat_buf);
-            call(self.emitter.rt.fd_filestat_get); drop;
-            local_get(stat_buf); i32_const(32); i32_add; i32_load(0); local_set(file_size);
-            local_get(file_size); call(self.emitter.rt.alloc); local_set(data_buf);
-            i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-            local_get(iov_ptr); local_get(data_buf); i32_store(0);
-            local_get(iov_ptr); local_get(file_size); i32_store(4);
-            i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-            local_get(opened_fd); local_get(iov_ptr); i32_const(1); local_get(nread_ptr);
-            call(self.emitter.rt.fd_read); drop;
-            local_get(opened_fd); call(self.emitter.rt.fd_close); drop;
-            local_get(nread_ptr); i32_load(0); local_set(file_size);
-        });
-
-        wasm!(self.func, {
-            local_get(file_size); i32_const(4); i32_add;
-            call(self.emitter.rt.alloc); local_set(str_ptr);
-            local_get(str_ptr); local_get(file_size); i32_store(0);
-        });
-
-        let counter = self.scratch.alloc_i32();
-        wasm!(self.func, {
-            i32_const(0); local_set(counter);
-            block_empty; loop_empty;
-            local_get(counter); local_get(file_size); i32_ge_u; br_if(1);
-            local_get(str_ptr); i32_const(4); i32_add; local_get(counter); i32_add;
-            local_get(data_buf); local_get(counter); i32_add;
-            i32_load8_u(0);
-            i32_store8(0);
-            local_get(counter); i32_const(1); i32_add; local_set(counter);
-            br(0);
-            end; end;
-        });
-        self.scratch.free_i32(counter);
-
-        wasm!(self.func, {
-            i32_const(8); call(self.emitter.rt.alloc); local_set(result_ptr);
-            local_get(result_ptr); i32_const(0); i32_store(0);
-            local_get(result_ptr); local_get(str_ptr); i32_store(4);
-            local_get(result_ptr);
-            end;
-        });
-
-        self.scratch.free_i32(errno);
-        self.scratch.free_i32(str_ptr);
-        self.scratch.free_i32(result_ptr);
-        self.scratch.free_i32(nread_ptr);
-        self.scratch.free_i32(iov_ptr);
-        self.scratch.free_i32(data_buf);
-        self.scratch.free_i32(file_size);
-        self.scratch.free_i32(stat_buf);
-        self.scratch.free_i32(opened_fd);
-        self.scratch.free_i32(fd_out_ptr);
-        self.scratch.free_i32(resolved_fd);
-        self.scratch.free_i32(path_len);
-        self.scratch.free_i32(path_ptr);
-        self.scratch.free_i32(path_str);
-    }
-
-    /// io module: print, read_line, read_all
-    fn emit_io_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "print" => {
-                // io.print(s: String) -> Unit
-                // Same as println but WITHOUT the trailing newline.
-                let s = self.scratch.alloc_i32();
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    local_set(s);
-                    // iov[0].buf = s + 4 (skip length prefix)
-                    i32_const(0);
-                    local_get(s); i32_const(4); i32_add;
-                    i32_store(0);
-                    // iov[0].len = *s (load length)
-                    i32_const(4);
-                    local_get(s); i32_load(0);
-                    i32_store(0);
-                    // fd_write(stdout=1, iovs=0, iovs_len=1, nwritten=8)
-                    i32_const(1); i32_const(0); i32_const(1); i32_const(8);
-                    call(self.emitter.rt.fd_write);
-                    drop;
-                });
-                self.scratch.free_i32(s);
-            }
-            "read_line" => {
-                // io.read_line() -> String
-                // Read one byte at a time from stdin (fd=0) until '\n' or EOF.
-                // Accumulate into a heap buffer, then build an Almide string.
-                let buf = self.scratch.alloc_i32();       // growing buffer ptr
-                let capacity = self.scratch.alloc_i32();  // current capacity
-                let len = self.scratch.alloc_i32();       // bytes read so far
-                let iov_ptr = self.scratch.alloc_i32();   // iov struct for fd_read
-                let nread_ptr = self.scratch.alloc_i32(); // nread output
-                let byte_buf = self.scratch.alloc_i32();  // 1-byte read target
-                let nread_val = self.scratch.alloc_i32(); // loaded nread value
-                let byte_val = self.scratch.alloc_i32();  // loaded byte value
-                let new_buf = self.scratch.alloc_i32();   // for realloc copy
-                let copy_i = self.scratch.alloc_i32();    // copy loop counter
-                let result = self.scratch.alloc_i32();    // final string ptr
-
-                // Initial capacity = 256
-                wasm!(self.func, {
-                    i32_const(256); call(self.emitter.rt.alloc); local_set(buf);
-                    i32_const(256); local_set(capacity);
-                    i32_const(0); local_set(len);
-                    // Allocate iov (8 bytes) and nread (4 bytes) and byte_buf (1 byte)
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                    i32_const(1); call(self.emitter.rt.alloc); local_set(byte_buf);
-                });
-
-                // Main read loop
-                wasm!(self.func, {
-                    block_empty; loop_empty;
-                });
-
-                // Grow buffer if full: len >= capacity
-                wasm!(self.func, {
-                    local_get(len); local_get(capacity); i32_ge_u;
-                    if_empty;
-                      // Double capacity
-                      local_get(capacity); i32_const(2); i32_mul; local_set(capacity);
-                      local_get(capacity); call(self.emitter.rt.alloc); local_set(new_buf);
-                      // Copy old data
-                      i32_const(0); local_set(copy_i);
-                      block_empty; loop_empty;
-                        local_get(copy_i); local_get(len); i32_ge_u; br_if(1);
-                        local_get(new_buf); local_get(copy_i); i32_add;
-                        local_get(buf); local_get(copy_i); i32_add; i32_load8_u(0);
-                        i32_store8(0);
-                        local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                        br(0);
-                      end; end;
-                      local_get(new_buf); local_set(buf);
-                    end;
-                });
-
-                // Set up iov to read 1 byte into byte_buf
-                wasm!(self.func, {
-                    local_get(iov_ptr); local_get(byte_buf); i32_store(0);
-                    local_get(iov_ptr); i32_const(1); i32_store(4);
-                    // fd_read(stdin=0, iov_ptr, 1, nread_ptr)
-                    i32_const(0);
-                    local_get(iov_ptr);
-                    i32_const(1);
-                    local_get(nread_ptr);
-                    call(self.emitter.rt.fd_read);
-                    drop;
-                });
-
-                // Check nread: if 0, EOF → break
-                wasm!(self.func, {
-                    local_get(nread_ptr); i32_load(0); local_set(nread_val);
-                    local_get(nread_val); i32_eqz;
-                    br_if(1); // break outer block
-                });
-
-                // Load byte, check for '\n'
-                wasm!(self.func, {
-                    local_get(byte_buf); i32_load8_u(0); local_set(byte_val);
-                    local_get(byte_val); i32_const(10); i32_eq; // '\n'
-                    br_if(1); // break outer block (don't include '\n' in result)
-                });
-
-                // Append byte to buffer
-                wasm!(self.func, {
-                    local_get(buf); local_get(len); i32_add;
-                    local_get(byte_val);
-                    i32_store8(0);
-                    local_get(len); i32_const(1); i32_add; local_set(len);
-                    br(0); // continue loop
-                    end; end; // end loop, end block
-                });
-
-                // Build Almide string [len:i32][data:u8...]
-                wasm!(self.func, {
-                    local_get(len); i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(result);
-                    local_get(result); local_get(len); i32_store(0);
-                    // Copy buf[0..len] to result+4
-                    i32_const(0); local_set(copy_i);
-                    block_empty; loop_empty;
-                      local_get(copy_i); local_get(len); i32_ge_u; br_if(1);
-                      local_get(result); i32_const(4); i32_add; local_get(copy_i); i32_add;
-                      local_get(buf); local_get(copy_i); i32_add; i32_load8_u(0);
-                      i32_store8(0);
-                      local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                      br(0);
-                    end; end;
-                    local_get(result);
-                });
-
-                self.scratch.free_i32(result);
-                self.scratch.free_i32(copy_i);
-                self.scratch.free_i32(new_buf);
-                self.scratch.free_i32(byte_val);
-                self.scratch.free_i32(nread_val);
-                self.scratch.free_i32(byte_buf);
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(len);
-                self.scratch.free_i32(capacity);
-                self.scratch.free_i32(buf);
-            }
-            "read_all" => {
-                // io.read_all() -> String
-                // Read all bytes from stdin (fd=0) until EOF.
-                // Strategy: read in chunks of 4096 bytes, grow buffer as needed.
-                let buf = self.scratch.alloc_i32();
-                let capacity = self.scratch.alloc_i32();
-                let len = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nread_ptr = self.scratch.alloc_i32();
-                let nread_val = self.scratch.alloc_i32();
-                let new_buf = self.scratch.alloc_i32();
-                let copy_i = self.scratch.alloc_i32();
-                let chunk_buf = self.scratch.alloc_i32();
-                let result = self.scratch.alloc_i32();
-
-                // Initial capacity = 4096
-                wasm!(self.func, {
-                    i32_const(4096); call(self.emitter.rt.alloc); local_set(buf);
-                    i32_const(4096); local_set(capacity);
-                    i32_const(0); local_set(len);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                });
-
-                // Read loop
-                wasm!(self.func, {
-                    block_empty; loop_empty;
-                });
-
-                // Ensure we have room for at least 4096 bytes
-                wasm!(self.func, {
-                    local_get(capacity); local_get(len); i32_sub;
-                    i32_const(4096); i32_lt_u;
-                    if_empty;
-                      // Double capacity
-                      local_get(capacity); i32_const(2); i32_mul; local_set(capacity);
-                      local_get(capacity); call(self.emitter.rt.alloc); local_set(new_buf);
-                      // Copy old data
-                      i32_const(0); local_set(copy_i);
-                      block_empty; loop_empty;
-                        local_get(copy_i); local_get(len); i32_ge_u; br_if(1);
-                        local_get(new_buf); local_get(copy_i); i32_add;
-                        local_get(buf); local_get(copy_i); i32_add; i32_load8_u(0);
-                        i32_store8(0);
-                        local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                        br(0);
-                      end; end;
-                      local_get(new_buf); local_set(buf);
-                    end;
-                });
-
-                // Read chunk into buf+len, up to (capacity - len) bytes
-                wasm!(self.func, {
-                    local_get(iov_ptr); local_get(buf); local_get(len); i32_add; i32_store(0);
-                    local_get(iov_ptr); local_get(capacity); local_get(len); i32_sub; i32_store(4);
-                    // fd_read(stdin=0, iov_ptr, 1, nread_ptr)
-                    i32_const(0);
-                    local_get(iov_ptr);
-                    i32_const(1);
-                    local_get(nread_ptr);
-                    call(self.emitter.rt.fd_read);
-                    drop;
-                });
-
-                // Check nread: if 0, EOF → break
-                wasm!(self.func, {
-                    local_get(nread_ptr); i32_load(0); local_set(nread_val);
-                    local_get(nread_val); i32_eqz;
-                    br_if(1);
-                    // Advance len
-                    local_get(len); local_get(nread_val); i32_add; local_set(len);
-                    br(0);
-                    end; end; // end loop, end block
-                });
-
-                // Build Almide string [len:i32][data:u8...]
-                wasm!(self.func, {
-                    local_get(len); i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(result);
-                    local_get(result); local_get(len); i32_store(0);
-                    // Copy buf[0..len] to result+4
-                    i32_const(0); local_set(copy_i);
-                    block_empty; loop_empty;
-                      local_get(copy_i); local_get(len); i32_ge_u; br_if(1);
-                      local_get(result); i32_const(4); i32_add; local_get(copy_i); i32_add;
-                      local_get(buf); local_get(copy_i); i32_add; i32_load8_u(0);
-                      i32_store8(0);
-                      local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                      br(0);
-                    end; end;
-                    local_get(result);
-                });
-
-                self.scratch.free_i32(result);
-                self.scratch.free_i32(chunk_buf);
-                self.scratch.free_i32(copy_i);
-                self.scratch.free_i32(new_buf);
-                self.scratch.free_i32(nread_val);
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(len);
-                self.scratch.free_i32(capacity);
-                self.scratch.free_i32(buf);
-            }
-            "read_byte" => {
-                // io.read_byte() -> Int
-                // Read 1 byte from stdin via fd_read. Return byte as i64, or -1i64 on EOF.
-                let byte_buf = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nread_ptr = self.scratch.alloc_i32();
-
-                wasm!(self.func, {
-                    i32_const(1); call(self.emitter.rt.alloc); local_set(byte_buf);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                    // iov: buf = byte_buf, len = 1
-                    local_get(iov_ptr); local_get(byte_buf); i32_store(0);
-                    local_get(iov_ptr); i32_const(1); i32_store(4);
-                    // fd_read(stdin=0, iov, 1, nread_ptr)
-                    i32_const(0); local_get(iov_ptr); i32_const(1); local_get(nread_ptr);
-                    call(self.emitter.rt.fd_read);
-                    drop;
-                    // if nread == 0 → -1i64, else byte as i64
-                    local_get(nread_ptr); i32_load(0); i32_eqz;
-                    if_i64;
-                      i64_const(-1);
-                    else_;
-                      local_get(byte_buf); i32_load8_u(0); i64_extend_i32_u;
-                    end;
-                });
-
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(byte_buf);
-            }
-            "read_n_bytes" => {
-                // io.read_n_bytes(n: Int) -> List[Int]
-                // Read up to n bytes from stdin, return as List[Int].
-                // List[Int] layout: [count:i32][elem0:i64][elem1:i64]...
-                let n = self.scratch.alloc_i32();
-                let raw_buf = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nread_ptr = self.scratch.alloc_i32();
-                let total = self.scratch.alloc_i32();
-                let nread_val = self.scratch.alloc_i32();
-                let list_ptr = self.scratch.alloc_i32();
-                let i = self.scratch.alloc_i32();
-
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    i32_wrap_i64; local_set(n);
-                    // Allocate raw read buffer of n bytes
-                    local_get(n); call(self.emitter.rt.alloc); local_set(raw_buf);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                    i32_const(0); local_set(total);
-                });
-
-                // Read loop: keep reading until total == n or EOF
-                wasm!(self.func, {
-                    block_empty; loop_empty;
-                      // Check if done
-                      local_get(total); local_get(n); i32_ge_u; br_if(1);
-                      // iov: buf = raw_buf + total, len = n - total
-                      local_get(iov_ptr); local_get(raw_buf); local_get(total); i32_add; i32_store(0);
-                      local_get(iov_ptr); local_get(n); local_get(total); i32_sub; i32_store(4);
-                      // fd_read(stdin=0, iov, 1, nread_ptr)
-                      i32_const(0); local_get(iov_ptr); i32_const(1); local_get(nread_ptr);
-                      call(self.emitter.rt.fd_read);
-                      drop;
-                      // Check nread
-                      local_get(nread_ptr); i32_load(0); local_set(nread_val);
-                      local_get(nread_val); i32_eqz; br_if(1); // EOF → break
-                      local_get(total); local_get(nread_val); i32_add; local_set(total);
-                      br(0);
-                    end; end;
-                });
-
-                // Build List[Int]: [total:i32][i64 * total]
-                wasm!(self.func, {
-                    // Allocate: 4 + total * 8
-                    local_get(total); i32_const(8); i32_mul; i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(list_ptr);
-                    local_get(list_ptr); local_get(total); i32_store(0);
-                    // Copy bytes → i64 elements
-                    i32_const(0); local_set(i);
-                    block_empty; loop_empty;
-                      local_get(i); local_get(total); i32_ge_u; br_if(1);
-                      local_get(list_ptr); i32_const(4); i32_add;
-                      local_get(i); i32_const(8); i32_mul; i32_add;
-                      local_get(raw_buf); local_get(i); i32_add; i32_load8_u(0); i64_extend_i32_u;
-                      i64_store(0);
-                      local_get(i); i32_const(1); i32_add; local_set(i);
-                      br(0);
-                    end; end;
-                    local_get(list_ptr);
-                });
-
-                self.scratch.free_i32(i);
-                self.scratch.free_i32(list_ptr);
-                self.scratch.free_i32(nread_val);
-                self.scratch.free_i32(total);
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(raw_buf);
-                self.scratch.free_i32(n);
-            }
-            "write" | "write_bytes" => {
-                // io.write(data: Bytes) — layout [len:i32][u8 data...], same as print
-                // io.write_bytes(data: List[Int]) — layout [len:i32][i64 elements...]
-                if func == "write" {
-                    let s = self.scratch.alloc_i32();
-                    self.emit_expr(&args[0]);
-                    wasm!(self.func, {
-                        local_set(s);
-                        i32_const(0);
-                        local_get(s); i32_const(4); i32_add;
-                        i32_store(0);
-                        i32_const(4);
-                        local_get(s); i32_load(0);
-                        i32_store(0);
-                        i32_const(1); i32_const(0); i32_const(1); i32_const(8);
-                        call(self.emitter.rt.fd_write);
-                        drop;
-                    });
-                    self.scratch.free_i32(s);
-                } else {
-                    // write_bytes: List[Int] → convert i64 to u8 then write
-                    let list_ptr = self.scratch.alloc_i32();
-                    let len = self.scratch.alloc_i32();
-                    let tmp_buf = self.scratch.alloc_i32();
-                    let i = self.scratch.alloc_i32();
-                    self.emit_expr(&args[0]);
-                    wasm!(self.func, {
-                        local_set(list_ptr);
-                        local_get(list_ptr); i32_load(0); local_set(len);
-                        local_get(len); call(self.emitter.rt.alloc); local_set(tmp_buf);
-                        i32_const(0); local_set(i);
-                        block_empty; loop_empty;
-                          local_get(i); local_get(len); i32_ge_u; br_if(1);
-                          local_get(tmp_buf); local_get(i); i32_add;
-                          local_get(list_ptr); i32_const(4); i32_add;
-                          local_get(i); i32_const(8); i32_mul; i32_add;
-                          i64_load(0); i32_wrap_i64;
-                          i32_store8(0);
-                          local_get(i); i32_const(1); i32_add; local_set(i);
-                          br(0);
-                        end; end;
-                        i32_const(0); local_get(tmp_buf); i32_store(0);
-                        i32_const(4); local_get(len); i32_store(0);
-                        i32_const(1); i32_const(0); i32_const(1); i32_const(8);
-                        call(self.emitter.rt.fd_write);
-                        drop;
-                    });
-                    self.scratch.free_i32(i);
-                    self.scratch.free_i32(tmp_buf);
-                    self.scratch.free_i32(len);
-                    self.scratch.free_i32(list_ptr);
-                }
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
-
-    /// process module: exit, stdin_lines
-    fn emit_process_call(&mut self, func: &str, args: &[IrExpr]) {
-        match func {
-            "exit" => {
-                // process.exit(code: Int) -> Unit
-                // Emit code arg (i64), wrap to i32, call proc_exit
-                self.emit_expr(&args[0]);
-                wasm!(self.func, {
-                    i32_wrap_i64;
-                    call(self.emitter.rt.proc_exit);
-                });
-            }
-            "stdin_lines" => {
-                // process.stdin_lines() -> List[String]
-                // Strategy: read all stdin, then split by '\n'.
-                // 1. Read all stdin into a raw buffer (same logic as io.read_all)
-                // 2. Split by '\n', building a list of Almide strings
-                let buf = self.scratch.alloc_i32();
-                let capacity = self.scratch.alloc_i32();
-                let len = self.scratch.alloc_i32();
-                let iov_ptr = self.scratch.alloc_i32();
-                let nread_ptr = self.scratch.alloc_i32();
-                let nread_val = self.scratch.alloc_i32();
-                let new_buf = self.scratch.alloc_i32();
-                let copy_i = self.scratch.alloc_i32();
-                let scan_i = self.scratch.alloc_i32();
-                let line_start = self.scratch.alloc_i32();
-                let line_len = self.scratch.alloc_i32();
-                let line_ptr = self.scratch.alloc_i32();
-                let list_ptr = self.scratch.alloc_i32();
-                let list_cap = self.scratch.alloc_i32();
-                let list_count = self.scratch.alloc_i32();
-                let new_list = self.scratch.alloc_i32();
-                let result = self.scratch.alloc_i32();
-
-                // --- Phase 1: read all stdin ---
-                wasm!(self.func, {
-                    i32_const(4096); call(self.emitter.rt.alloc); local_set(buf);
-                    i32_const(4096); local_set(capacity);
-                    i32_const(0); local_set(len);
-                    i32_const(8); call(self.emitter.rt.alloc); local_set(iov_ptr);
-                    i32_const(4); call(self.emitter.rt.alloc); local_set(nread_ptr);
-                });
-
-                wasm!(self.func, {
-                    block_empty; loop_empty;
-                });
-
-                // Grow if needed
-                wasm!(self.func, {
-                    local_get(capacity); local_get(len); i32_sub;
-                    i32_const(4096); i32_lt_u;
-                    if_empty;
-                      local_get(capacity); i32_const(2); i32_mul; local_set(capacity);
-                      local_get(capacity); call(self.emitter.rt.alloc); local_set(new_buf);
-                      i32_const(0); local_set(copy_i);
-                      block_empty; loop_empty;
-                        local_get(copy_i); local_get(len); i32_ge_u; br_if(1);
-                        local_get(new_buf); local_get(copy_i); i32_add;
-                        local_get(buf); local_get(copy_i); i32_add; i32_load8_u(0);
-                        i32_store8(0);
-                        local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                        br(0);
-                      end; end;
-                      local_get(new_buf); local_set(buf);
-                    end;
-                });
-
-                // Read chunk
-                wasm!(self.func, {
-                    local_get(iov_ptr); local_get(buf); local_get(len); i32_add; i32_store(0);
-                    local_get(iov_ptr); local_get(capacity); local_get(len); i32_sub; i32_store(4);
-                    i32_const(0);
-                    local_get(iov_ptr);
-                    i32_const(1);
-                    local_get(nread_ptr);
-                    call(self.emitter.rt.fd_read);
-                    drop;
-                });
-
-                wasm!(self.func, {
-                    local_get(nread_ptr); i32_load(0); local_set(nread_val);
-                    local_get(nread_val); i32_eqz;
-                    br_if(1);
-                    local_get(len); local_get(nread_val); i32_add; local_set(len);
-                    br(0);
-                    end; end;
-                });
-
-                // --- Phase 2: split buf[0..len] by '\n' into List[String] ---
-                // List layout: [count:i32][elem0:i32][elem1:i32]...
-                // Each elem is a ptr to Almide String [len:i32][data:u8...]
-                // We'll build with a growable array of i32 pointers.
-                wasm!(self.func, {
-                    // Initial list capacity: 64 elements (i32 ptrs)
-                    i32_const(64); local_set(list_cap);
-                    local_get(list_cap); i32_const(4); i32_mul;
-                    call(self.emitter.rt.alloc); local_set(list_ptr);
-                    i32_const(0); local_set(list_count);
-                    i32_const(0); local_set(scan_i);
-                    i32_const(0); local_set(line_start);
-                });
-
-                // Scan loop: iterate through buf looking for '\n'
-                wasm!(self.func, {
-                    block_empty; loop_empty;
-                      local_get(scan_i); local_get(len); i32_ge_u;
-                      br_if(1);
-                });
-
-                // Check if buf[scan_i] == '\n'
-                wasm!(self.func, {
-                      local_get(buf); local_get(scan_i); i32_add; i32_load8_u(0);
-                      i32_const(10); i32_eq;
-                      if_empty;
-                });
-
-                // Found '\n': build string from line_start..scan_i
-                wasm!(self.func, {
-                        local_get(scan_i); local_get(line_start); i32_sub; local_set(line_len);
-                        // Allocate Almide string
-                        local_get(line_len); i32_const(4); i32_add;
-                        call(self.emitter.rt.alloc); local_set(line_ptr);
-                        local_get(line_ptr); local_get(line_len); i32_store(0);
-                        // Copy line data
-                        i32_const(0); local_set(copy_i);
-                        block_empty; loop_empty;
-                          local_get(copy_i); local_get(line_len); i32_ge_u; br_if(1);
-                          local_get(line_ptr); i32_const(4); i32_add; local_get(copy_i); i32_add;
-                          local_get(buf); local_get(line_start); i32_add; local_get(copy_i); i32_add;
-                          i32_load8_u(0);
-                          i32_store8(0);
-                          local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                          br(0);
-                        end; end;
-                });
-
-                // Grow list if needed
-                wasm!(self.func, {
-                        local_get(list_count); local_get(list_cap); i32_ge_u;
-                        if_empty;
-                          local_get(list_cap); i32_const(2); i32_mul; local_set(list_cap);
-                          local_get(list_cap); i32_const(4); i32_mul;
-                          call(self.emitter.rt.alloc); local_set(new_list);
-                          // Copy old list ptrs
-                          i32_const(0); local_set(copy_i);
-                          block_empty; loop_empty;
-                            local_get(copy_i); local_get(list_count); i32_ge_u; br_if(1);
-                            local_get(new_list); local_get(copy_i); i32_const(4); i32_mul; i32_add;
-                            local_get(list_ptr); local_get(copy_i); i32_const(4); i32_mul; i32_add;
-                            i32_load(0);
-                            i32_store(0);
-                            local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                            br(0);
-                          end; end;
-                          local_get(new_list); local_set(list_ptr);
-                        end;
-                });
-
-                // Append line_ptr to list
-                wasm!(self.func, {
-                        local_get(list_ptr); local_get(list_count); i32_const(4); i32_mul; i32_add;
-                        local_get(line_ptr); i32_store(0);
-                        local_get(list_count); i32_const(1); i32_add; local_set(list_count);
-                        // line_start = scan_i + 1
-                        local_get(scan_i); i32_const(1); i32_add; local_set(line_start);
-                      end; // end if '\n'
-                });
-
-                // Advance scan_i
-                wasm!(self.func, {
-                      local_get(scan_i); i32_const(1); i32_add; local_set(scan_i);
-                      br(0);
-                    end; end; // end loop, end block
-                });
-
-                // Handle last line (if no trailing '\n')
-                wasm!(self.func, {
-                    local_get(line_start); local_get(len); i32_lt_u;
-                    if_empty;
-                      local_get(len); local_get(line_start); i32_sub; local_set(line_len);
-                      local_get(line_len); i32_const(4); i32_add;
-                      call(self.emitter.rt.alloc); local_set(line_ptr);
-                      local_get(line_ptr); local_get(line_len); i32_store(0);
-                      i32_const(0); local_set(copy_i);
-                      block_empty; loop_empty;
-                        local_get(copy_i); local_get(line_len); i32_ge_u; br_if(1);
-                        local_get(line_ptr); i32_const(4); i32_add; local_get(copy_i); i32_add;
-                        local_get(buf); local_get(line_start); i32_add; local_get(copy_i); i32_add;
-                        i32_load8_u(0);
-                        i32_store8(0);
-                        local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                        br(0);
-                      end; end;
-                });
-
-                // Grow list if needed for last line
-                wasm!(self.func, {
-                      local_get(list_count); local_get(list_cap); i32_ge_u;
-                      if_empty;
-                        local_get(list_cap); i32_const(2); i32_mul; local_set(list_cap);
-                        local_get(list_cap); i32_const(4); i32_mul;
-                        call(self.emitter.rt.alloc); local_set(new_list);
-                        i32_const(0); local_set(copy_i);
-                        block_empty; loop_empty;
-                          local_get(copy_i); local_get(list_count); i32_ge_u; br_if(1);
-                          local_get(new_list); local_get(copy_i); i32_const(4); i32_mul; i32_add;
-                          local_get(list_ptr); local_get(copy_i); i32_const(4); i32_mul; i32_add;
-                          i32_load(0);
-                          i32_store(0);
-                          local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                          br(0);
-                        end; end;
-                        local_get(new_list); local_set(list_ptr);
-                      end;
-                      // Append last line
-                      local_get(list_ptr); local_get(list_count); i32_const(4); i32_mul; i32_add;
-                      local_get(line_ptr); i32_store(0);
-                      local_get(list_count); i32_const(1); i32_add; local_set(list_count);
-                    end; // end if line_start < len
-                });
-
-                // Build final Almide List: [count:i32][elem0:i32][elem1:i32]...
-                // elem_size = 4 (i32 pointer)
-                wasm!(self.func, {
-                    local_get(list_count); i32_const(4); i32_mul; i32_const(4); i32_add;
-                    call(self.emitter.rt.alloc); local_set(result);
-                    local_get(result); local_get(list_count); i32_store(0);
-                    // Copy list_ptr[0..list_count] to result+4
-                    i32_const(0); local_set(copy_i);
-                    block_empty; loop_empty;
-                      local_get(copy_i); local_get(list_count); i32_ge_u; br_if(1);
-                      local_get(result); i32_const(4); i32_add;
-                      local_get(copy_i); i32_const(4); i32_mul; i32_add;
-                      local_get(list_ptr); local_get(copy_i); i32_const(4); i32_mul; i32_add;
-                      i32_load(0);
-                      i32_store(0);
-                      local_get(copy_i); i32_const(1); i32_add; local_set(copy_i);
-                      br(0);
-                    end; end;
-                    local_get(result);
-                });
-
-                self.scratch.free_i32(result);
-                self.scratch.free_i32(new_list);
-                self.scratch.free_i32(list_count);
-                self.scratch.free_i32(list_cap);
-                self.scratch.free_i32(list_ptr);
-                self.scratch.free_i32(line_ptr);
-                self.scratch.free_i32(line_len);
-                self.scratch.free_i32(line_start);
-                self.scratch.free_i32(scan_i);
-                self.scratch.free_i32(copy_i);
-                self.scratch.free_i32(new_buf);
-                self.scratch.free_i32(nread_val);
-                self.scratch.free_i32(nread_ptr);
-                self.scratch.free_i32(iov_ptr);
-                self.scratch.free_i32(len);
-                self.scratch.free_i32(capacity);
-                self.scratch.free_i32(buf);
-            }
-            _ => {
-                self.emit_stub_call(args);
-            }
-        }
-    }
 }
